% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-load.R
\name{color}
\alias{color}
\title{Generates a number of most distinctive colors for PCA.}
\usage{
color()
}
\description{
This function should not be used directly,
only inside \code{\link{initialize_format}} function.
}
\examples{
\dontrun{
color()
}

}
\keyword{internal}
