% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.biplot_title}
\alias{.biplot_title}
\title{Generate title for PCA plots}
\usage{
.biplot_title(sample_type)
}
\arguments{
\item{sample_type}{sample type such as "All", "Healthy Tissue (GTEx)",
"Primary Tumor (TCGA)", or "Metastatic Tumor (TCGA)"}
}
\value{
a string for title
}
\description{
This function generates title for PCA plots based on the input value.
It should not be used directly, only inside \code{\link[=.biplot]{.biplot()}} or
\code{\link[=.selected_biplot]{.selected_biplot()}} function.
}
\seealso{
Other helper function for PCA plotting: 
\code{\link{.biplot}()},
\code{\link{.selected_biplot}()}
}
\concept{helper function for PCA plotting}
\keyword{internal}
