% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.CNV.all.cancer}
\alias{.CNV.all.cancer}
\title{Calculates the frequency of CNV status in all TCGA cancer types combined}
\usage{
.CNV.all.cancer(df)
}
\arguments{
\item{df}{\code{TCGA.CNV.sampletype.subset} generated inside \code{\link{plot.bargraph.CNV.TCGA}}}
}
\value{
a summary table ranking the EIF4F gene by the frequencies of duplication (CNV threshold labeled as “1” in the dataset)
}
\description{
This function calculates the frequency of each CNV status across tumors in all TCGA cancer types for every EIF4F gene.
}
\details{
It should not be used directly, only inside \code{\link{plot.bargraph.CNV.TCGA}} function.
}
\examples{
\dontrun{.CNV.all.cancer(TCGA.CNV.sampletype.subset)}
}
\keyword{internal}
