% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.CNV.barplot}
\alias{.CNV.barplot}
\title{Stacked bar plots of the CNV status}
\usage{
.CNV.barplot(df)
}
\arguments{
\item{df}{\code{EIF.CNV.ind.cancer} generated inside \code{\link{plot.bargraph.CNV.TCGA}}}
}
\value{
stacked bar plots for CNV status of each gene in an individual cancer type.
}
\description{
This function generates stacked bar plots for CNV status of each gene in an individual cancer type.
}
\details{
This plot function uses dataset generated from \code{\link{.CNV.ind.cancer}} function
It should not be used directly, only inside \code{\link{plot.bargraph.CNV.TCGA}} function.
}
\examples{
\dontrun{lapply(EIF.CNV.ind.cancer, .CNV.barplot)}
}
\keyword{internal}
