% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.CNV.ind.cancer}
\alias{.CNV.ind.cancer}
\title{Calculates the frequency of CNV status in individual TCGA cancer types}
\usage{
.CNV.ind.cancer(df, x)
}
\arguments{
\item{df}{\code{TCGA.CNV.sampletype.subset} generated inside \code{\link{plot.bargraph.CNV.TCGA}}}

\item{x}{one gene from the input argument of \code{\link{plot.bargraph.CNV.TCGA}}}
}
\value{
a list with the summary table of CNV in individual TCGA cancer types and gene name
}
\description{
a data analysis function that calculates the frequency of CNV status for each EIF4F gene in individual TCGA cancer types.
}
\details{
It should not be used directly, only inside \code{\link{plot.bargraph.CNV.TCGA}} function.
}
\examples{
\dontrun{lapply(EIF,.CNV.ind.cancer,df = TCGA.CNV.sampletype.subset)}
}
\keyword{internal}
