% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.CNV.sum.barplot}
\alias{.CNV.sum.barplot}
\title{Stacked bar plots of the CNV status summary}
\usage{
.CNV.sum.barplot(data)
}
\arguments{
\item{data}{summary table from \code{.CNV.all.cancer(TCGA.CNV.sampletype.subset)}}
}
\value{
stacked bar plots showing the summary table of the CNV statuses
}
\description{
This function generates the stacked bar plots to the summary of CNV statuses
}
\details{
This plot function uses dataset generated from \code{\link{.CNV.all.cancer}} function

It should not be used directly, only inside \code{\link{plot.bargraph.CNV.TCGA}}function.
}
\examples{
\dontrun{.CNV.sum.barplot(.CNV.all.cancer(TCGA.CNV.sampletype.subset))}
}
\keyword{internal}
