% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.CNV_all_cancer}
\alias{.CNV_all_cancer}
\title{Calculates the frequency of CNV status in all TCGA cancer types combined}
\usage{
.CNV_all_cancer(df)
}
\arguments{
\item{df}{\code{.TCGA_CNV_sampletype_subset} generated inside \code{\link{.plot_bargraph_CNV_TCGA}}}
}
\value{
a summary table ranking the EIF4F gene by the frequencies of duplication (CNV threshold labeled as “1” in the dataset)
}
\description{
This function calculates the frequency of each CNV status across tumors in all TCGA cancer types for every EIF4F gene.
}
\details{
It should not be used directly, only inside \code{\link{.plot_bargraph_CNV_TCGA}} function.
}
\examples{
\dontrun{
.CNV_all_cancer(.TCGA_CNV_sampletype_subset)
}
}
\keyword{internal}
