% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.CNV_barplot}
\alias{.CNV_barplot}
\title{Stacked bar plots of the CNV status}
\usage{
.CNV_barplot(df)
}
\arguments{
\item{df}{\code{.EIF_CNV_ind_cancer} generated inside
\code{\link[=.plot_bargraph_CNV_TCGA]{.plot_bargraph_CNV_TCGA()}}}
}
\description{
This function generates stacked bar plots for CNV status of
each gene in an individual cancer type.
}
\details{
This plot function uses dataset generated from
\code{\link[=.CNV_ind_cancer]{.CNV_ind_cancer()}} function

It should not be used directly, only inside
\code{\link[=.plot_bargraph_CNV_TCGA]{.plot_bargraph_CNV_TCGA()}} function.

Side effect: stacked bar plots on screen and as saved pdf files to show
CNV status of each gene in an individual cancer type.
}
\examples{
\dontrun{
lapply(.EIF_CNV_ind_cancer, .CNV_barplot)
}

}
\seealso{
Other helper function for CNV data plotting: 
\code{\link{.CNV_sum_barplot}()},
\code{\link{.CNVratio_boxplot}()},
\code{\link{.matrix_plot}()}
}
\concept{helper function for CNV data plotting}
\keyword{internal}
