% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.CNV_ind_cancer}
\alias{.CNV_ind_cancer}
\title{Calculates the frequency of CNV status in individual TCGA cancer types}
\usage{
.CNV_ind_cancer(df, x)
}
\arguments{
\item{df}{\code{.TCGA_CNV_sampletype_subset} generated inside
\code{\link[=.plot_bargraph_CNV_TCGA]{.plot_bargraph_CNV_TCGA()}}}

\item{x}{one gene from the input argument of
\code{\link[=.plot_bargraph_CNV_TCGA]{.plot_bargraph_CNV_TCGA()}}}
}
\value{
a list with the summary table of CNV in individual
TCGA cancer types and gene name
}
\description{
a data analysis function that calculates the frequency of
CNV status for each EIF4F gene in individual TCGA cancer types.
}
\details{
It should not be used directly, only inside
\code{\link[=.plot_bargraph_CNV_TCGA]{.plot_bargraph_CNV_TCGA()}} function.
}
\examples{
\dontrun{
lapply(EIF, .CNV_ind_cancer, df = .TCGA_CNV_sampletype_subset)
}

}
\keyword{internal}
