% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.CNV_sum_barplot}
\alias{.CNV_sum_barplot}
\title{Stacked bar plots of the CNV status summary}
\usage{
.CNV_sum_barplot(data)
}
\arguments{
\item{data}{summary table from
\code{.CNV_all_cancer(.TCGA_CNV_sampletype_subset)}}
}
\value{
stacked bar plots showing the summary table of the CNV statuses
}
\description{
This function generates the stacked bar plots to the summary
of CNV statuses
}
\details{
This plot function uses dataset generated from
\code{\link[=.CNV_all_cancer]{.CNV_all_cancer()}} function

It should not be used directly, only inside
\code{\link[=.plot_bargraph_CNV_TCGA]{.plot_bargraph_CNV_TCGA()}}function.
}
\examples{
\dontrun{
.CNV_sum_barplot(.CNV_all_cancer(.TCGA_CNV_sampletype_subset))
}

}
\keyword{internal}
