% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.CNVratio.boxplot}
\alias{.CNVratio.boxplot}
\title{Box plots of the CNV ratios in tumors vs adjacent normals}
\usage{
.CNVratio.boxplot(df)
}
\arguments{
\item{df}{\code{EIF.CNVratio.ind.cancer} generated inside \code{\link{plot.boxgraph.CNVratio.TCGA}}}
}
\value{
boxplot for CNV ratios in tumors vs adjacent normals from individual TCGA cancer types.
}
\description{
This function generates boxplot for CNV ratios in tumors vs adjacent normals from individual TCGA cancer types.
}
\details{
This plot function uses dataset \code{TCGA.CNVratio.sampletype.subset} generated from \code{\link{plot.boxgraph.CNVratio.TCGA}} function.

It should not be used directly, only inside \code{\link{plot.boxgraph.CNVratio.TCGA}} function.
}
\examples{
\dontrun{lapply(EIF.CNVratio.ind.cancer, .CNVratio.boxplot)}
}
\keyword{internal}
