% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.CNVratio.tumor}
\alias{.CNVratio.tumor}
\title{Calculates the frequency of CNV status in all TCGA cancer types combined}
\usage{
.CNVratio.tumor(df, x)
}
\arguments{
\item{df}{\code{TCGA.CNVratio.sampletype.subset} generated inside \code{\link{plot.boxgraph.CNVratio.TCGA}}}

\item{x}{one gene from the input argument of \code{\link{plot.boxgraph.CNVratio.TCGA}}}
}
\value{
a list with the data frame of CNV ratio of the input gene in individual TCGA cancer types and gene name
}
\description{
This function selects the CNV ratio data in tumors vs adjacent normals from individual TCGA cancer types for each input gene.
}
\details{
It should not be used directly, only inside \code{\link{plot.boxgraph.CNVratio.TCGA}} function.
}
\examples{
\dontrun{lapply(EIF,.CNVratio.tumor,df = TCGA.CNVratio.sampletype.subset)}
}
\keyword{internal}
