% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.CNVratio_boxplot}
\alias{.CNVratio_boxplot}
\title{Box plots of the CNV ratios in tumors vs adjacent normals}
\usage{
.CNVratio_boxplot(df)
}
\arguments{
\item{df}{\code{.EIF_CNVratio_ind_cancer} generated
inside \code{\link[=.plot_boxgraph_CNVratio_TCGA]{.plot_boxgraph_CNVratio_TCGA()}}}
}
\description{
This function generates boxplot for CNV ratios in tumors
vs adjacent normals from individual TCGA cancer types.
}
\details{
This plot function uses dataset \code{.TCGA_CNVratio_sampletype_subset}
generated from \code{\link[=.plot_boxgraph_CNVratio_TCGA]{.plot_boxgraph_CNVratio_TCGA()}} function.
It should not be used directly, only inside \code{\link[=.plot_boxgraph_CNVratio_TCGA]{.plot_boxgraph_CNVratio_TCGA()}}
function.

Side effects:

(1) boxplot on screen and as saved pdf files to show CNV ratios
in tumors vs adjacent normals from individual TCGA cancer types.
}
\examples{
\dontrun{
lapply(.EIF_CNVratio_ind_cancer, .CNVratio_boxplot)
}

}
\seealso{
Other helper function for CNV data plotting: 
\code{\link{.CNV_barplot}()},
\code{\link{.CNV_sum_barplot}()},
\code{\link{.matrix_plot}()}
}
\concept{helper function for CNV data plotting}
\keyword{internal}
