% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.combine_CORs_list}
\alias{.combine_CORs_list}
\title{Combine all the EIF4F correlating gene data from tumor and healthy
tissues}
\usage{
.combine_CORs_list(df1, df2)
}
\arguments{
\item{df1}{\code{cor_value_combined} of \code{EIF.cor.tumor[[2]]}}

\item{df2}{\code{cor_value_combined} of \code{EIF.cor.normal[[2]]}}
}
\value{
data frame with posCORs or negCORs from both tumors and healthy tissue samples
}
\description{
This function
\itemize{
\item combines all the EIF4F correlating genes and their correlation
coefficients to a data frame, with the data generated from
\code{\link[=.EIF_correlation]{.EIF_correlation()}}.
\item selects significant correlating genes with \code{\link[=.is_significant_correlation]{.is_significant_correlation()}}
}

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}
function.
}
\examples{
\dontrun{
.combine_COR_list(df1 = EIF.cor.tumor[[1]], df2 = EIF.cor.normal[[1]])
}

}
\seealso{
Other helper function for correlation analysis: 
\code{\link{.combine_CORs_summary}()},
\code{\link{.get_cluster_genes}()},
\code{\link{.is_significant_correlation}()}
}
\concept{helper function for correlation analysis}
\keyword{internal}
