% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.combine_CORs_summary}
\alias{.combine_CORs_summary}
\title{Combine the summary of number of posCORs or negCORs from sample types}
\usage{
.combine_CORs_summary(COR_tbl1, sample_type1, COR_tbl2, sample_type2)
}
\arguments{
\item{COR_tbl1}{a COR summary table of specific sample type \code{EIF.cor.tumor[[2]]}}

\item{sample_type1}{the label of sample type where \code{COR_df1} is derived from}

\item{COR_tbl2}{a COR summary table of specific sample type \code{EIF.cor.normal[[2]]}}

\item{sample_type2}{the label of sample type where \code{COR_df2} is derived from}
}
\value{
a data frame with combined COR summary table
}
\description{
This function combines the summary of number of posCORs or negCORs from
sample types, using the CORs table generated from \code{\link[=.EIF_correlation]{.EIF_correlation()}}.

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}
function.
}
\examples{
\dontrun{
.combine_CORs_summary(
COR_tbl1 = EIF.cor.tumor[[2]], sample_type1 = "tumor",
COR_tbl2 = EIF.cor.normal[[2]], sample_type2 = "normal")
}

}
\seealso{
Other helper function for correlation analysis: 
\code{\link{.combine_CORs_list}()},
\code{\link{.get_cluster_genes}()},
\code{\link{.is_significant_correlation}()}
}
\concept{helper function for correlation analysis}
\keyword{internal}
