% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.correlation_coefficient}
\alias{.correlation_coefficient}
\title{Calculate the correlation co-efficiency between two gene expressions}
\usage{
.correlation_coefficient(gene1, gene2, df)
}
\arguments{
\item{gene1}{gene name}

\item{gene2}{gene name}

\item{df}{expression dataset}
}
\value{
a dataframe with correlation coefficiency between two gene expression
}
\description{
A helper function calculates the correlation co-efficiency between two gene
expressions, using the RNAseq expression dataset.

It should not be used directly, only inside \code{\link[=.EIF_correlation]{.EIF_correlation()}}
function.
}
\seealso{
Other helper function to identify correlating genes for EIF4F genes: 
\code{\link{.EIF_correlation}()},
\code{\link{.is_significant_negcor}()},
\code{\link{.is_significant_poscor}()},
\code{\link{.summarize_counts}()}
}
\concept{helper function to identify correlating genes for EIF4F genes}
\keyword{internal}
