% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-Correlation.R
\name{.Cors_bargraph}
\alias{.Cors_bargraph}
\title{Plot bargraph for the numbers of correlating genes}
\usage{
.Cors_bargraph(df, x, CORs, coord_flip.ylim)
}
\arguments{
\item{df}{correlation data}

\item{x}{input argument of \code{\link{.plot_Corr_RNAseq_TCGA_GTEX}}}

\item{CORs}{posCOR or negCORs for the title of Venn diagram}

\item{coord_flip.ylim}{the limit of y axis in the bar plot}
}
\value{
bargraph for the numbers of posCOR or negCORs
}
\description{
This function draw bargraph, using the correlation data
generated from \code{\link{.EIF_correlation}}.

It should not be used directly, only inside \code{\link{.plot_Corr_RNAseq_TCGA_GTEX}} function.
}
\examples{
\dontrun{
.Cors_bargraph(
  df = EIF.cor, x = x,
  CORs = "posCORs", coord_flip.ylim = 14000
)
}
}
\keyword{internal}
