% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.CORs_coeff_heatmap}
\alias{.CORs_coeff_heatmap}
\title{Visualize associations between different sources of correlation data
and reveal potential pattern}
\usage{
.CORs_coeff_heatmap(df, tissue_type)
}
\arguments{
\item{df}{combined COR data generated from \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}}

\item{tissue_type}{"All" for all tissue type, or "Lung" for specific tissue}
}
\value{
a data structure that includes clustering analysis results.
}
\description{
This function draw heatmap of combined COR list data generated from
\code{\link[=.EIF_correlation]{.EIF_correlation()}}.

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}
function.

Side effects:

(1) heatmap on screen and as pdf file to show correlation strength
and clustering pattern of EIF4F correlating genes.
}
\examples{
\dontrun{
.CORs_coeff_heatmap(df = DF, tissue_type = tissue_type)
}

}
\seealso{
Other helper function for correlation analysis plotting: 
\code{\link{.CORs_summary_bargraph}()},
\code{\link{.CORs_vennplot}()},
\code{\link{.pathway_dotplot}()}
}
\concept{helper function for correlation analysis plotting}
\keyword{internal}
