% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.CORs_summary_bargraph}
\alias{.CORs_summary_bargraph}
\title{Plot bargraph for the numbers of correlating genes}
\usage{
.CORs_summary_bargraph(df, tissue_type, CORs_type, coord_flip.ylim)
}
\arguments{
\item{df}{combined CORs summary table}

\item{tissue_type}{input argument of \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}}

\item{CORs_type}{posCORs or negCORs for the title of Venn diagram}

\item{coord_flip.ylim}{the limit of y axis in the bar plot}
}
\value{
bargraph for the numbers of posCOR or negCORs
}
\description{
This function draw bargraph, using the correlation data generated from
\code{\link[=.EIF_correlation]{.EIF_correlation()}}.

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}
function.

Side effects:

(1) bar graphs on screen and as pdf file to show the numbers of
identified correlating genes for each EIF4F subunit
}
\examples{
\dontrun{
.CORs_summary_bargraph(
 df = EIF.cor,
 tissue_type = tissue_type,
 CORs_type = "posCORs",
 coord_flip.ylim = 14000)
}

}
\seealso{
Other helper function for correlation analysis plotting: 
\code{\link{.CORs_coeff_heatmap}()},
\code{\link{.CORs_vennplot}()},
\code{\link{.pathway_dotplot}()}
}
\concept{helper function for correlation analysis plotting}
\keyword{internal}
