% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.Cors_vennplot}
\alias{.Cors_vennplot}
\title{\subsection{Correlation analysis and plotting}{

Plot Venn diagrams for correlating genes
}}
\usage{
.Cors_vennplot(df, x, z, CORs)
}
\arguments{
\item{df}{correlation data}

\item{x}{input argument of \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}}

\item{z}{tumor or normal for the title of Venn diagram}

\item{CORs}{posCOR or negCORs for the title of Venn diagram}
}
\value{
vennDiagram for posCOR or negCORs
}
\description{
This function draw Venn diagrams, using the correlation data generated from
\code{\link[=.EIF_correlation]{.EIF_correlation()}}.

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}
function.
}
\examples{
\dontrun{
.Cors_vennplot(df = EIF.cor.tumor[[3]], x = x, z = "tumor", CORs = "posCOR")
}

}
\keyword{internal}
