% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.count_CORs_tumor_normal}
\alias{.count_CORs_tumor_normal}
\title{Count the numbers of correlating genes}
\usage{
.count_CORs_tumor_normal(df1, df2)
}
\arguments{
\item{df1}{\code{EIF.cor.tumor[[2]]}}

\item{df2}{\code{EIF.cor.normal[[2]]}}
}
\value{
bargraph for posCOR or negCOR numbers
}
\description{
This function counts the numbers of correlating genes, using the data
generated from \code{\link[=.EIF_correlation]{.EIF_correlation()}}.

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}
function.
}
\examples{
\dontrun{
.count_CORs_tumor_normal(
  df1 = EIF.cor.tumor[[2]],
  df2 = EIF.cor.normal[[2]]
)
}

}
\keyword{internal}
