% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.EIF_cor_normal_tumor}
\alias{.EIF_cor_normal_tumor}
\title{Combine all the EIF4F correlating gene data from tumor and healthy tissues}
\usage{
.EIF_cor_normal_tumor(df1, df2)
}
\arguments{
\item{df1}{\code{EIF.cor.tumor[[2]]}}

\item{df2}{\code{EIF.cor.normal[[2]]}}
}
\value{
data frame with posCOR or negCOR from both tumors and healthy tissue samples
}
\description{
This function combines all the EIF4F correlating genes and their correlation
coefficients to a data frame, with the data generated from
\code{\link[=.EIF_correlation]{.EIF_correlation()}}.

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}
function.
}
\examples{
\dontrun{
.EIF_cor_normal_tumor(df1 = EIF.cor.tumor[[1]], df2 = EIF.cor.normal[[1]])
}

}
\keyword{internal}
