% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-Survival.R
\name{.forest_graph}
\alias{.forest_graph}
\title{Forest plots of COX-PH results}
\usage{
.forest_graph(data, output.file, plot.title, x.tics, x.range)
}
\arguments{
\item{data}{output dataset generated from \code{\link{.univariable_analysis}} or \code{\link{.multivariable_analysis}} function}

\item{output.file}{output file name}

\item{plot.title}{title name of the forest graph}

\item{x.tics}{tics for x axis}

\item{x.range}{range for x axis}
}
\description{
This function should not be used directly, only inside \code{\link{.plot_CoxPH_RNAseq_TCGA}} function.
}
\keyword{internal}
