% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.get.TCGA.CNV.ratio}
\alias{.get.TCGA.CNV.ratio}
\title{Read CNV ratio dataset from TCGA}
\usage{
.get.TCGA.CNV.ratio()
}
\value{
a data frame that contains CNV ratio data for each tumor
}
\description{
This function reads the CNV ratio data "broad.mit.edu_PANCAN_Genome_Wide_SNP_6_whitelisted.gene.xena",
which calculates the ratios of CNV value in each tumor to the average CNV value from normals in the same cancer type.
}
\details{
The function also removes possible duplicated tumor samples and samples with NAs in the dataset.

It should not be used directly, only inside \code{\link{initialize.cnv.data}} function.
}
\examples{
\dontrun{.get.TCGA.CNV.ratio()}
}
