% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.get.TCGA.CNV}
\alias{.get.TCGA.CNV}
\title{Read CNV threshold dataset from TCGA}
\usage{
.get.TCGA.CNV()
}
\value{
a data frame that contains CNV threshold data for all TCGA tumor
}
\description{
This function reads the threshold CNV data "Gistic2_CopyNumber_Gistic2_all_thresholded.by_genes",
which groups the CNV values in tumors into five categories.
}
\details{
The function also removes possible duplicated tumor samples and samples with NAs in the dataset.

It should not be used directly, only inside \code{\link{initialize.cnv.data}} function.
}
\examples{
\dontrun{.get.TCGA.CNV()}
}
\keyword{internal}
