% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.get.TCGA.CNV.value}
\alias{.get.TCGA.CNV.value}
\title{Read CNV value dataset from TCGA}
\usage{
.get.TCGA.CNV.value()
}
\value{
a data frame that contains CNV value data for each tumor
}
\description{
This function reads the unthreshold CNV value data from TCGA "Gistic2_CopyNumber_Gistic2_all_data_by_genes".
}
\details{
The function also removes possible duplicated tumor samples and samples with NAs in the dataset.

It should not be used directly, only inside \code{\link{initialize.cnv.data}} function.
}
\examples{
\dontrun{.get.TCGA.CNV.value()}
}
\keyword{internal}
