% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAProCorr.R
\name{.get_CCLE_RNAseq_subset}
\alias{.get_CCLE_RNAseq_subset}
\title{Select subset of CCLE RNAseq data}
\usage{
.get_CCLE_RNAseq_subset(gene_name)
}
\arguments{
\item{gene_name}{gene name}
}
\value{
a data frame of CCLE RNAseq data from the input \code{gene_name} genes
}
\description{
A helper function selects the CCLE RNAseq data of the input gene
\code{EIF}.
}
\details{
The function should not be used directly, only inside
\code{\link[=.plot_scatter_RNApro_CCLE]{.plot_scatter_RNApro_CCLE()}} function.
}
\examples{
\dontrun{
.get_CCLE_RNAseq_subset()
}

}
\seealso{
Other helper function for RNA protein correlation analysis: 
\code{\link{.RNApro_scatterplot}()},
\code{\link{.get_CCLE_Proteomics_subset}()}
}
\concept{helper function for RNA protein correlation analysis}
\keyword{internal}
