% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.get_cluster_genes}
\alias{.get_cluster_genes}
\title{Retrieve the gene names from each cluster in heatmap}
\usage{
.get_cluster_genes(df1, df2)
}
\arguments{
\item{df1}{combined COR data generated from \code{\link[=.combine_CORs_list]{.combine_CORs_list()}}}

\item{df2}{gene name in the order of heatmap generated from
\code{\link[=.CORs_coeff_heatmap]{.CORs_coeff_heatmap()}}}
}
\value{
a heatmap plot and dataframe with gene names of clustering
}
\description{
This function retrieve the gene names from the clusters in heatmap from
\code{\link[=.CORs_coeff_heatmap]{.CORs_coeff_heatmap()}}.

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}
function.
}
\examples{
\dontrun{
.get_cluster_genes(df1 = DF, df2 = ht1)
}

}
\seealso{
Other helper function for correlation analysis: 
\code{\link{.combine_CORs_list}()},
\code{\link{.combine_CORs_summary}()},
\code{\link{.is_significant_correlation}()}
}
\concept{helper function for correlation analysis}
\keyword{internal}
