% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteomics.R
\name{.get_CPTAC_LUAD_Phosproteomics_subset}
\alias{.get_CPTAC_LUAD_Phosproteomics_subset}
\title{Select subset of CPTAC phosproteomics data}
\usage{
.get_CPTAC_LUAD_Phosproteomics_subset(protein_list)
}
\arguments{
\item{protein_list}{protein name, passed \code{protein_list} argument from
\code{\link[=.plot_boxgraph_protein_CPTAC]{.plot_boxgraph_protein_CPTAC()}}}
}
\value{
a data frame of CPTAC LUAD  data from the input \code{protein_list} genes
}
\description{
This function selected the CPTAC LUAD phosphoproteomics data from the
input \code{protein_list}.
}
\details{
The function should not be used directly, only inside
\code{\link[=.plot_boxgraph_protein_CPTAC]{.plot_boxgraph_protein_CPTAC()}} function.
}
\examples{
\dontrun{
.get_CCLE_RNAseq_subset()
}

}
\seealso{
Other helper function for protein coexpression analysis: 
\code{\link{.get_CPTAC_LUAD_Proteomics_subset}()},
\code{\link{.protein_boxplot}()},
\code{\link{.protein_scatterplot}()}
}
\concept{helper function for protein coexpression analysis}
\keyword{internal}
