% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.get_TCGA_CNV}
\alias{.get_TCGA_CNV}
\title{Read CNV threshold dataset from TCGA}
\usage{
.get_TCGA_CNV()
}
\value{
a data frame that contains CNV threshold data for all TCGA tumor
}
\description{
This function reads the threshold CNV data
\code{Gistic2_CopyNumber_Gistic2_all_thresholded.by_genes}, which grouped the
CNV values with thresholds 3+, 3, 2, 1, 0, to represent high-level copy
number gain, low-level copy number gain, diploid, shallow deletion, or
deep deletion.
}
\details{
The function also removes possible duplicated tumor samples and
samples with NAs in the dataset.

It should not be used directly, only inside \code{\link[=initialize_cnv_data]{initialize_cnv_data()}}
function.
}
\examples{
\dontrun{
.get_TCGA_CNV()
}

}
\seealso{
Other helper function for data initialization: 
\code{\link{.get_TCGA_CNV_ratio}()},
\code{\link{.get_TCGA_CNV_value}()},
\code{\link{.get_TCGA_GTEX_RNAseq}()},
\code{\link{.get_TCGA_RNAseq}()}
}
\concept{helper function for data initialization}
\keyword{internal}
