% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.get_TCGA_CNV_ratio}
\alias{.get_TCGA_CNV_ratio}
\title{Read CNV ratio dataset from TCGA}
\usage{
.get_TCGA_CNV_ratio()
}
\value{
a data frame that contains CNV ratio data for each tumor
}
\description{
This function reads the CNV ratio data
\code{broad.mit.edu_PANCAN_Genome_Wide_SNP_6_whitelisted.gene.xena}, in which
CNV ratios were calculated by dividing, for each gene, the estimated
gene-level CNV values in malignant tumors to the average CNV value in
normal adjacent tissues (NATs) of the same cancer type.
}
\details{
The function also removes possible duplicated tumor samples and
samples with NAs in the dataset.

It should not be used directly, only inside \code{\link[=initialize_cnv_data]{initialize_cnv_data()}}
function.
}
\examples{
\dontrun{
.get_TCGA_CNV_ratio()
}

}
\seealso{
Other helper function for data initialization: 
\code{\link{.get_TCGA_CNV_value}()},
\code{\link{.get_TCGA_CNV}()},
\code{\link{.get_TCGA_GTEX_RNAseq}()},
\code{\link{.get_TCGA_RNAseq}()}
}
\concept{helper function for data initialization}
\keyword{internal}
