% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.get_TCGA_CNV_value}
\alias{.get_TCGA_CNV_value}
\title{Read CNV value dataset from TCGA}
\usage{
.get_TCGA_CNV_value()
}
\value{
a data frame that contains CNV value data for each tumor
}
\description{
This function reads the unthreshold CNV value data from TCGA
\code{Gistic2_CopyNumber_Gistic2_all_data_by_genes}.
}
\details{
The function also removes possible duplicated tumor samples and
samples with NAs in the dataset.

It should not be used directly, only inside \code{\link[=initialize_cnv_data]{initialize_cnv_data()}}
function.
}
\examples{
\dontrun{
.get_TCGA_CNV_value()
}

}
\seealso{
Other helper function for data initialization: 
\code{\link{.get_TCGA_CNV_ratio}()},
\code{\link{.get_TCGA_CNV}()},
\code{\link{.get_TCGA_GTEX_RNAseq}()},
\code{\link{.get_TCGA_RNAseq}()}
}
\concept{helper function for data initialization}
\keyword{internal}
