% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{.get_TCGA_RNAseq}
\alias{.get_TCGA_RNAseq}
\title{Read the RNAseq data from TCGA}
\usage{
.get_TCGA_RNAseq()
}
\value{
a data frame that contains the RNAseq data from TCGA
}
\description{
A helper function reads the RNAseq data from TCGA
\code{EB++AdjustPANCAN_IlluminaHiSeq_RNASeqV2.geneExp.xena}.
}
\details{
This function removes possible duplicated tumor samples and samples with
NAs in the dataset.
It should not be used directly, only inside \code{\link[=initialize_survival_data]{initialize_survival_data()}}
function.
}
\examples{
\dontrun{
.get_TCGA_RNAseq()
}

}
\seealso{
Other helper function for data initialization: 
\code{\link{.get_TCGA_CNV_ratio}()},
\code{\link{.get_TCGA_CNV_value}()},
\code{\link{.get_TCGA_CNV}()},
\code{\link{.get_TCGA_GTEX_RNAseq}()}
}
\concept{helper function for data initialization}
\keyword{internal}
