% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.is_significant_correlation}
\alias{.is_significant_correlation}
\title{Select both positive and negative correlating genes based on
coefficiency and pvalue}
\usage{
.is_significant_correlation(magnitude, pvalue)
}
\arguments{
\item{magnitude}{correlation coefficiency value}

\item{pvalue}{p value for Pearson correlation}
}
\value{
a data frame of both positive and negative correlating gene with
correlation coefficiency and p value
}
\description{
This function selects both positive and negative correlating genes based on
coefficiency and pvalue.

It should not be used directly, only inside \code{\link[=.combine_CORs_list]{.combine_CORs_list()}}
function.
}
\seealso{
Other helper function for correlation analysis: 
\code{\link{.combine_CORs_list}()},
\code{\link{.combine_CORs_summary}()},
\code{\link{.get_cluster_genes}()}
}
\concept{helper function for correlation analysis}
\keyword{internal}
