% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.is_significant_poscor}
\alias{.is_significant_poscor}
\title{Select positive correlating genes based on coefficiency and pvalue}
\usage{
.is_significant_poscor(magnitude, pvalue)
}
\arguments{
\item{magnitude}{correlation coefficiency value}

\item{pvalue}{p value for Pearson correlation}
}
\value{
a data frame of positive correlating gene with correlation
coefficiency and p value
}
\description{
A helper function selects positive correlating genes based on coefficiency
and pvalue.

It should not be used directly, only inside \code{\link[=.EIF_correlation]{.EIF_correlation()}}
function.
}
\seealso{
Other helper function to identify correlating genes for EIF4F genes: 
\code{\link{.EIF_correlation}()},
\code{\link{.correlation_coefficient}()},
\code{\link{.is_significant_negcor}()},
\code{\link{.summarize_counts}()}
}
\concept{helper function to identify correlating genes for EIF4F genes}
\keyword{internal}
