% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{.KM_curve}
\alias{.KM_curve}
\title{Kaplan Meier survival analyses of gene expression}
\usage{
.KM_curve(gene, data, cutoff, tumor)
}
\arguments{
\item{gene}{gene name, passed \code{EIF} argument from \code{\link[=.plot_KM_RNAseq_TCGA]{.plot_KM_RNAseq_TCGA()}}}

\item{data}{\code{df} generated inside \code{\link[=.plot_KM_RNAseq_TCGA]{.plot_KM_RNAseq_TCGA()}}}

\item{cutoff}{percentage of gene expression}

\item{tumor}{all tumor types or specific type}
}
\value{
a KM plot
}
\description{
A helper function correlates the gene expression within tumor samples with
patient overall survival time from TCGA study groups
}
\details{
It should not be used directly, only inside \code{\link[=.plot_KM_RNAseq_TCGA]{.plot_KM_RNAseq_TCGA()}}
function.

Side effects:
(1) KM curve plots for TCGA patients with gene expression in their
tumors
}
\examples{
\dontrun{
.KM_curve(gene = EIF, data = df, cutoff = cutoff, tumor = tumor)
}

}
\seealso{
Other helper function for survival analysis: 
\code{\link{.multivariable_analysis}()},
\code{\link{.univariable_analysis}()}
}
\concept{helper function for survival analysis}
\keyword{internal}
