% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.matrix.plot}
\alias{.matrix.plot}
\title{Correlation coefficients for CNV values}
\usage{
.matrix.plot(df)
}
\arguments{
\item{df}{output of \code{TCGA.CNV.value \%>\% select(all_of(EIF))} generated inside \code{\link{plot.matrix.CNVcorr.TCGA}}}
}
\value{
stacked bar plots for CNV status of each gene in an individual cancer type.
}
\description{
This function calculates the correlation coefficients between every two genes and plot the correlation matrix with the function.
}
\details{
This plot function uses dataset \code{TCGA.CNV.value} generated from \code{\link{initialize.cnv.data}} function
It should not be used directly, only inside \code{\link{plot.matrix.CNVcorr.TCGA}} function.
}
\examples{
\dontrun{TCGA.CNV.value \%>\% dplyr::select(all_of(EIF)) \%>\% .matrix.plot()}
}
\keyword{internal}
