% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{.multivariable_analysis}
\alias{.multivariable_analysis}
\title{Multivariable Cox-PH analyses of gene expression}
\usage{
.multivariable_analysis(df, covariate_names)
}
\arguments{
\item{covariate_names}{gene names from the input argument of
\code{\link[=.plot_CoxPH_RNAseq_TCGA]{.plot_CoxPH_RNAseq_TCGA()}}}

\item{gene}{gene names, passed \code{EIF.list} argument from
\code{\link[=.plot_CoxPH_RNAseq_TCGA]{.plot_CoxPH_RNAseq_TCGA()}}}

\item{data}{\code{df1} generated inside \code{\link[=.plot_CoxPH_RNAseq_TCGA]{.plot_CoxPH_RNAseq_TCGA()}}}
}
\value{
a table of multivariable Cox-PH
}
\description{
This function generates univariable regression model of the gene expression
within tumor samples and patient overall survival time TCGA.
}
\details{
It should not be used directly, only inside
\code{\link[=.plot_CoxPH_RNAseq_TCGA]{.plot_CoxPH_RNAseq_TCGA()}} function.
}
\examples{
\dontrun{
.univariable_analysis(df = df1, covariate_names = EIF)
}

}
\keyword{internal}
