% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.pathway_dotplot}
\alias{.pathway_dotplot}
\title{Plot the enriched pathways from gene lists}
\usage{
.pathway_dotplot(df, pathway, tissue_type)
}
\arguments{
\item{df}{pathway enrichment analysis results generated from
\code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}}

\item{pathway}{pathway name, "GO", ""REACTOME", "KEGG"}

\item{tissue_type}{tissue type, the same argument of
\code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}}
}
\description{
This function plots pathway enrichment analysis results of the gene from each
cluster in heatmap generated from \code{\link[=.get_cluster_genes]{.get_cluster_genes()}}.

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}
function.

Side effects:

(1) dotplot on screen and as pdf file to show the enriched pathways
in the genes from different clusters in the heatmap
}
\examples{
\dontrun{
.pathway_dotplot(df = ck.REACTOME, pathway = "REACTOME",
 tissue_type = tissue_type)
}

}
\seealso{
Other helper function for correlation analysis plotting: 
\code{\link{.CORs_coeff_heatmap}()},
\code{\link{.CORs_summary_bargraph}()},
\code{\link{.CORs_vennplot}()}
}
\concept{helper function for correlation analysis plotting}
\keyword{internal}
