% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.plot_bargraph_CNV_TCGA}
\alias{.plot_bargraph_CNV_TCGA}
\title{Summary of CNV statuses in bar plots}
\usage{
.plot_bargraph_CNV_TCGA(gene_list)
}
\arguments{
\item{gene_list}{gene names in a vector of characters}
}
\description{
Provides the summary of CNV statuses of EIF4F genes in tumors from all TCGA
cancer types combined and in tumors from individual TCGA cancer types
}
\details{
This function
\itemize{
\item selects CNV and sample type data from only
EIF4F gene in the data frame \code{TCGA_CNV_sampletype} prepared
from \code{\link[=initialize_cnv_data]{initialize_cnv_data()}}.
\item uses the subset data \code{.TCGA_CNV_sampletype_subset} to perform the
CNV status analysis of all inquired genes with the functions
\code{\link[=.CNV_all_cancer]{.CNV_all_cancer()}} and plot the results as a bar plot
with \code{\link[=.CNV_sum_barplot]{.CNV_sum_barplot()}}
\item uses the same subset data to perform CNV status analysis for each
gene across all tumors types with the function \code{\link[=.CNV_ind_cancer]{.CNV_ind_cancer()}}
and plots the CNV results and \code{\link[=.CNV_barplot]{.CNV_barplot()}}
}

It should not be used directly, only inside \code{\link[=EIF4F_CNV_analysis]{EIF4F_CNV_analysis()}} function.

Side effects:

(1) the stacked bar plots on screen and as pdf file to show the summary
table of the CNV statuses of all \code{gene_list} in TCGA tumors

(2) stacked bar plots on screen and as saved pdf files to show
CNV status of each gene in \code{gene_list} from an individual cancer type
}
\examples{
\dontrun{
plot_bargraph_CNV_TCGA(c(
  "TP53", "EIF4A1", "EIF4A2", "EIF4E",
  "EIF4E2", "EIF4E3", "MYC", "EIF3D", "EIF4EBP1", "EIF4G1", "EIF4G2",
  "EIF4G3", "PABPC1", "MKNK1", "MKNK2"
))
}

}
\seealso{
Other composite function to call CNV data analysis and plotting: 
\code{\link{.plot_boxgraph_CNVratio_TCGA}()}
}
\concept{composite function to call CNV data analysis and plotting}
\keyword{internal}
