% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.plot_boxgraph_CNVratio_TCGA}
\alias{.plot_boxgraph_CNVratio_TCGA}
\title{CNV ratios in tumors vs adjacent normal tissues across tumor types}
\usage{
.plot_boxgraph_CNVratio_TCGA(EIF_list)
}
\arguments{
\item{EIF_list}{gene names in a vector of characters}
}
\value{
stacked bar plots for grouped CNV status of \code{EIF_list} in TCGA tumors
}
\description{
This function generates boxplot for CNV ratios in tumors vs adjacent normals from individual TCGA cancer types.
}
\details{
This function first selects CNV and sample type data from only EIF4F gene
in the data frame \code{TCGA_CNVratio_sampletype} prepared from \code{\link{initialize_cnv_data}}.

Then it uses the subset data \code{.TCGA_CNVratio_sampletype_subset} to perform CNV status analysis for each gene across
all tumors types with the function \code{\link{.CNVratio_tumor}} and plots the CNV results and \code{\link{.CNVratio_boxplot}}

It should not be used directly, only inside \code{\link{EIF4F_CNV_analysis}} function.
}
\examples{
\dontrun{
plot_boxgraph_CNVratio_TCGA(c("EIF4A1", "EIF4E", "EIF4EBP1", "EIF4G1"))
}
}
\keyword{internal}
