% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteomics.R
\name{.plot_boxgraph_protein_CPTAC}
\alias{.plot_boxgraph_protein_CPTAC}
\title{Comparison of protein and phosphorylation levels among different
stages of LUAD}
\usage{
.plot_boxgraph_protein_CPTAC(protein_list)
}
\arguments{
\item{protein_list}{protein names in a vector of string}
}
\description{
A composite function analyzes differential expression of eIF4F protein and
phosphorylation in CPTAC LUAD data and plots results as boxplots
}
\details{
This function
\itemize{
\item merges the LUAD proteomics values from EIF4F genes
in the data frame prepared from \code{\link[=.get_CPTAC_LUAD_Proteomics_subset]{.get_CPTAC_LUAD_Proteomics_subset()}} and
phosphoproteomics data of the same protein in the data frame prepared
from \code{\link[=.get_CPTAC_LUAD_Phosproteomics_subset]{.get_CPTAC_LUAD_Phosproteomics_subset()}}.
\item uses the combined data to compare the abundance across different
tumor stages, and plot the results with the function \code{\link[=.protein_boxplot]{.protein_boxplot()}}
}

This function is not accessible to the user and will not show at the users'
workspace. It can only be called by the exported \code{\link[=EIF4F_Proteomics_analysis]{EIF4F_Proteomics_analysis()}}
function.

Side effects:

(1) boxplot to show the different expression of (phospho)protein
across different clinical tumor stages
}
\examples{
\dontrun{
.plot_boxgraph_protein_CPTAC(c(
  "EIF4G1", "EIF4A1", "EIF4E", "EIF4EBP1",
  "AKT1", "MTOR", "EIF4B", "EIF4H", "MKNK1", "MKNK2"
))
}

}
\seealso{
Other composite function to call protein coexpression and differential
expression analysis: 
\code{\link{.plot_scatterplot_protein_CPTAC}()}
}
\concept{composite function to call protein coexpression and differential
expression analysis}
\keyword{internal}
