% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{.plot_boxgraph_RNAseq_TCGA}
\alias{.plot_boxgraph_RNAseq_TCGA}
\title{Compare the expression of EIF4F genes}
\usage{
.plot_boxgraph_RNAseq_TCGA(gene_list)
}
\arguments{
\item{gene_list}{gene names in a vector of characters}
}
\description{
This function generates a summary box plot to compare the expression of all
EIF4F genes across TCGA cancer types, box plots for differential gene
expression in tumors tumors vs adjacent normal tissues for each  gene.
and a violin plot to compare differential gene expression in primary,
metastatic tumors vs adjacent normal tissues from all TCGA cancer types
combined.
}
\details{
This function
\itemize{
\item selects RNASeq of input genes and columns including \code{sample.type},
\code{primary.disease}, and \code{primary.site},from the data frame
\code{TCGA_GTEX_RNAseq_sampletype} that was prepared by the function
\code{\link[=initialize_RNAseq_data]{initialize_RNAseq_data()}} to produce a subset data frame called
\code{.TCGA_GTEX_RNAseq_sampletype_subset}
\item with the subset data, calls the functions \code{\link[=.RNAseq_all_gene]{.RNAseq_all_gene()}} to compare
relative mRNA expression of all inquired genes and calls the functions
\code{\link[=.RNAseq_grouped_boxplot]{.RNAseq_grouped_boxplot()}} plot the results as a grouped box plot.
\item performs differential expression analysis for each gene across all
tumors types with the function \code{\link[=.RNAseq_ind_gene]{.RNAseq_ind_gene()}} and plots with
\code{\link[=.RNAseq_boxplot]{.RNAseq_boxplot()}}.
\item performs differential expression analysis for each gene in
primary, metastatic tumors vs adjacent normal tissues from all TCGA cancer
types combined with the function \code{\link[=.RNAseq_tumortype]{.RNAseq_tumortype()}} and plots with
\code{\link[=.violinplot]{.violinplot()}}.
}

It should not be used directly, only inside \code{\link[=EIF4F_DEG_analysis]{EIF4F_DEG_analysis()}} function.

Side effects:

(1) box plots on screen and as pdf file to compare relative gene
expression in across all TCGA cancer types

(2) box plots on screen and as pdf file to show differential
gene expression in tumors vs NATs in TCGA cancer types

(3) violin plots on screen and as pdf file to show
differential gene expression in primary, metastatic tumors vs
adjacent normal tissues from all combined TCGA cancer types
}
\examples{
\dontrun{
.plot_boxgraph_RNAseq_TCGA(c(
  "EIF4G1", "EIF4G2", "EIF4G3",
  "PABPC1", "EIF4A1", "EIF4A2", "EIF4B", "EIF4H", "EIF4E",
  "EIF4E2", "EIF4E3", "EIF4EBP1", "EIF3D"
))
}

}
\concept{composite function to call DEG gene analysis and plotting}
\keyword{internal}
