% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.plot_Corr_RNAseq_TCGA_GTEX}
\alias{.plot_Corr_RNAseq_TCGA_GTEX}
\title{\subsection{Master functions to analyze the EIF4F CORs}{

Perform correlation analysis on RNAseq data from all tumors and healthy
tissues
}}
\usage{
.plot_Corr_RNAseq_TCGA_GTEX(x)
}
\arguments{
\item{x}{all tissue/tumor types or one specific type}
}
\value{
Venn digram, bargraph, heatmap, and dotplot
}
\description{
This function
\itemize{
\item finds correlating genes of EIF4F from \code{TCGA_GTEX_RNAseq_sampletype}
RNAseq data generated by \code{\link[=initialize_RNAseq_data]{initialize_RNAseq_data()}}.
\item identifies positively or negatively correlating genes from tumor or
healthy samples,
\item finds the overlap of CORs for EIF4F subunits with Vennplot,
\item compares the number of CORs for EIF4F subunits with bargraph,
\item visualizes the correlation strength with heatmap,
\item founds enriched pathways within correlating genes
}

This function should not be used directly, only inside
\code{\link[=EIF4F_Corrgene_analysis]{EIF4F_Corrgene_analysis()}} function.
}
\examples{
\dontrun{
.plot_Corr_RNAseq_TCGA_GTEX(x = "All")
.plot_Corr_RNAseq_TCGA_GTEX(x = "Lung")
}

}
\keyword{internal}
