% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-Survival.R
\name{.plot_CoxPH_RNAseq_TCGA}
\alias{.plot_CoxPH_RNAseq_TCGA}
\title{Survival analyses of TCGA patients with expression of \code{EIF} in their tumors by Cox-PH method}
\usage{
.plot_CoxPH_RNAseq_TCGA(EIF.list, tumor)
}
\arguments{
\item{EIF.list}{gene names in a vector of characters}

\item{tumor}{all tumor types or specific type}
}
\value{
forest graph showing the relation between survival of TCGA patients and expression of \code{EIF} in their tumors
}
\description{
This function makes regression models between the gene expression within tumor samples and patient overall survival time.
}
\details{
This function first selects RNAseq of the query gene,
survival data and cancer types from the dataset \code{TCGA_RNAseq_OS_sampletype} prepared from \code{\link{initialize_survival_data}}.

With the subset data \code{df1}, it makes univariable regression models with \code{\link{.univariable_analysis}}
and multivariable regression models with \code{\link{.multivariable_analysis}}.

It plots the results as a forest graph with \code{\link{.forest_graph}}

This function should not be used directly, only inside \code{\link{EIF4F_Survival_analysis}} function.
}
\examples{
\dontrun{
.plot_CoxPH_RNAseq_TCGA(c(
  "EIF4E", "EIF4E2", "EIF4E3",
  "EIF4G1", "EIF4G2", "EIF4G3", "EIF4A1", "EIF4A2", "EIF3D", "EIF3E", "EIF4EBP1",
  "EIF4EBP2", "MKNK1", "MKNK2", "EIF4B", "EIF4H", "MTOR", "MYC"
), "All")
}
\dontrun{
.plot_CoxPH_RNAseq_TCGA(c(
  "EIF4E", "EIF4E2", "EIF4E3",
  "EIF4G1", "EIF4G2", "EIF4G3", "EIF4A1", "EIF4A2", "EIF3D", "EIF3E", "EIF4EBP1",
  "EIF4EBP2", "MKNK1", "MKNK2", "EIF4B", "EIF4H", "MTOR", "MYC"
), "lung adenocarcinoma")
}
}
\keyword{internal}
