% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{.plot_KM_RNAseq_TCGA}
\alias{.plot_KM_RNAseq_TCGA}
\title{\subsection{Master functions to call survival analysis and plotting}{

Survival analyses of TCGA patients with expression in their tumors by
Kaplan Meier method
}}
\usage{
.plot_KM_RNAseq_TCGA(EIF, cutoff, tumor)
}
\arguments{
\item{EIF}{gene name}

\item{cutoff}{percentage of gene expression for patient stratification}

\item{tumor}{all tumor types or specific type}
}
\value{
KM curve plots for TCGA patients with expression of \code{EIF} in their tumors
}
\description{
This function generates a Kaplan Meier plot to compare the expression of
one gene in TCGA cancer types.
}
\details{
This function
\itemize{
\item selects RNAseq of the query gene, survival data and cancer types from
the dataset \code{TCGA_RNAseq_OS_sampletype} prepared from
\code{\link[=initialize_survival_data]{initialize_survival_data()}}.
\item compares the survival data from patients with top or bottom percents of
gene expression, and plot the results as a KM curve plot by \code{\link[=.KM_curve]{.KM_curve()}}.
}

This function should not be used directly, only inside
\code{\link[=EIF4F_Survival_analysis]{EIF4F_Survival_analysis()}} function.
}
\examples{
\dontrun{
plot.km.EIF.tumor(EIF = "EIF4E", cutoff = 0.2, tumor = "lung adenocarcinoma")
}

}
\keyword{internal}
