% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.plot_matrix_CNVcorr_TCGA}
\alias{.plot_matrix_CNVcorr_TCGA}
\title{Correlation matrix for CNV values}
\usage{
.plot_matrix_CNVcorr_TCGA(EIF_list)
}
\arguments{
\item{EIF_list}{gene names in a vector of characters}
}
\value{
the correlation matrix plot for \code{EIF_list} CNV values in TCGA tumors
}
\description{
generates correlation matrix for eIF4F CNV in tumors from all TCGA cancer type combined
}
\details{
This function first selects unthreshold CNV values from only EIF4F gene
in the data frame \code{TCGA_CNV_value} prepared from \code{\link{initialize_cnv_data}}.

Then it uses the subset data to calculate the correlation coefficients
between every two genes and plot the correlation matrix with the function \code{\link{.matrix_plot}}

It should not be used directly, only inside \code{\link{EIF4F_CNV_analysis}} function.
}
\examples{
\dontrun{
plot_matrix_CNVcorr_TCGA(c("EIF4A1", "EIF4E", "EIF4EBP1", "EIF4G1"))
}
}
\keyword{internal}
