% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.plot_PCA_CPTAC_LUAD}
\alias{.plot_PCA_CPTAC_LUAD}
\title{Perform PCA on proteomics data from CPATC LUADs and matched healthy lung
tissues}
\usage{
.plot_PCA_CPTAC_LUAD(EIF.list)
}
\arguments{
\item{EIF.list}{gene names in a vector of characters}
}
\value{
PCA bioplot, matrix, scree plot
}
\description{
This function performs standard PCA on selected proteomics data from
\code{CPTAC_LUAD_Proteomics} prepared by \code{\link[=initialize_proteomics_data]{initialize_proteomics_data()}}.

It perform PCA on combined tumors from CPTAC LUAD and and
matched healthy lung tissues.

This function should not be used directly, only inside \code{\link[=EIF4F_PCA]{EIF4F_PCA()}} function.
}
\examples{
\dontrun{
.plot_PCA_TCGA_GTEX_tumor(c(
  "EIF4G1", "EIF4A1", "EIF4E",
  "EIF4EBP1", "PABPC1", "MKNK1", "MKNK2"
), "Lung")
}

}
\keyword{internal}
