% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.plot_PCA_TCGA_GTEX}
\alias{.plot_PCA_TCGA_GTEX}
\title{Perform PCA on RNAseq data from all tumors and healthy tissues}
\usage{
.plot_PCA_TCGA_GTEX(gene_list)
}
\arguments{
\item{gene_list}{gene names in a vector of characters}
}
\description{
A composite function call PCA on RNAseq data from all tumors
and healthy tissues and plot results
}
\details{
This function
\itemize{
\item selects RNAseq data of \code{gene_list} genes of specific sample types from
\code{TCGA_GTEX_RNAseq_sampletype} by \code{\link[=.get_df_subset]{.get_df_subset()}}.
\item performs three PCAs by \code{\link[=.RNAseq_PCA]{.RNAseq_PCA()}} on \enumerate{
\item tumors samples from all TCGA cancer types
\item healthy tissue samples from all GTEx healthy tissue types
\item TCGA tumors and GTEx healthy tissue samples combined}
\item generates biplot, screen plot and matrix plots by \code{\link[=.biplot]{.biplot()}}
\item generates subset biplots for PCA results of combined TCGA tumors and
GTEx healthy tissues with \code{\link[=.selected_biplot]{.selected_biplot()}} function
}

This function should not be used directly, only inside \code{\link[=EIF4F_PCA]{EIF4F_PCA()}}
function.

Side effects:

(1) PCA biplots (PCA score plot + loading plot) on screen and as pdf files:
PCA score plot shows the clusters of samples based on their similarity and
loading plot shows how strongly each characteristic influences a principal
component.

(2) matrix plots on screen and as pdf files to show the quality of
representation of the variables.

(3) scree plots on screen and as pdf files to display how much variation
each principal component captures from the data.
}
\examples{
\dontrun{
.plot_PCA_TCGA_GTEX(c(
  "EIF4E", "EIF4G1", "EIF4A1",
  "EIF4EBP1", "PABPC1", "MKNK1", "MKNK2"
))
}

}
\seealso{
Other composite function to call PCA and plot results: 
\code{\link{.plot_PCA_CPTAC_LUAD}()},
\code{\link{.plot_PCA_TCGA_GTEX_tumor}()}
}
\concept{composite function to call PCA and plot results}
\keyword{internal}
