% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-PCA.R
\name{.plot_PCA_TCGA_GTEX_tumor}
\alias{.plot_PCA_TCGA_GTEX_tumor}
\title{Perform PCA on RNAseq data from one tumor type and matched healthy tissue}
\usage{
.plot_PCA_TCGA_GTEX_tumor(EIF.list, tissue)
}
\arguments{
\item{EIF.list}{gene names in a vector of characters}

\item{tissue}{tissue type}
}
\value{
PCA bioplot, matrix, scree plot
}
\description{
This function performs standard PCA on selected RNAseq data from
\code{EIF.list} genes generated by \code{\link{.get_df_subset}}.

It perform PCA on combined tumors from TCGA cancer types and GTEx healthy tissues with the same \code{tissue} of origin.

This function should not be used directly, only inside \code{\link{EIF4F_PCA}} function.
}
\examples{
\dontrun{
.plot_PCA_TCGA_GTEX_tumor(c(
  "EIF4G1", "EIF4A1", "EIF4E",
  "EIF4EBP1", "PABPC1", "MKNK1", "MKNK2"
), "Lung")
}

}
\keyword{internal}
