% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAProCorr.R
\name{.plot_scatter_RNApro_LUAD}
\alias{.plot_scatter_RNApro_LUAD}
\title{Correlation between LUAD RNAseq and proteomics data}
\usage{
.plot_scatter_RNApro_LUAD(gene_name)
}
\arguments{
\item{gene_name}{gene name}
}
\description{
A composite function generates correlation scatter plot for eIF4F RNAseq and
proteomics data from LUAD.
}
\details{
This function
\itemize{
\item merges the LUAD RNAseq data of EIF4F genes prepared from
\code{CPTAC_LUAD_Proteomics} and proteomics data of the EIF4F proteins
prepared from \code{CPTAC_LUAD_RNAseq}.
\item uses the combined data to calculate the correlation coefficients between
protein and RNA levels, and plot the result with the function
\code{\link[=.RNApro_scatterplot]{.RNApro_scatterplot()}}
}

This function is not accessible to the user and will not show at the users'
workspace. It can only be called by the exported \code{\link[=EIF4F_RNA_pro_correlation]{EIF4F_RNA_pro_correlation()}}
function.

Side effects:

(1) scatter plot to show correlation between RNA and protein
expressions of input \code{gene_name} in CPTAC LUAD samples
}
\examples{
\dontrun{
lapply(c("EIF4G1", "EIF4A1", "EIF4E"), .plot_scatter_RNApro_LUAD)
}

}
\seealso{
Other composite function to call RNA protein correlation analysis and
plotting: 
\code{\link{.plot_scatter_RNApro_CCLE}()}
}
\concept{composite function to call RNA protein correlation analysis and
plotting}
\keyword{internal}
