% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteomics.R
\name{.protein_boxplot}
\alias{.protein_boxplot}
\title{Boxplots of phosphor-proteomics data across clinic stages}
\usage{
.protein_boxplot(df, protein_name)
}
\arguments{
\item{df}{a data frame of the combined proteomics and phosphorylation data
generated inside \code{\link[=.plot_boxgraph_protein_CPTAC]{.plot_boxgraph_protein_CPTAC()}}}

\item{protein_name}{protein name}
}
\description{
A helper function draws boxplots for protein differential
expression across clinical stages
}
\details{
This function should not be used directly,
only inside \code{\link[=.plot_boxgraph_protein_CPTAC]{.plot_boxgraph_protein_CPTAC()}}.

Side effects:

(1) boxplots to show the different expression of protein across
different clinical tumor stages
}
\seealso{
Other helper function for protein coexpression analysis: 
\code{\link{.get_CPTAC_LUAD_Phosproteomics_subset}()},
\code{\link{.get_CPTAC_LUAD_Proteomics_subset}()},
\code{\link{.protein_scatterplot}()}
}
\concept{helper function for protein coexpression analysis}
\keyword{internal}
