% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.protein_imputePCA}
\alias{.protein_imputePCA}
\title{Perform imputPCA on proteomics data}
\usage{
.protein_imputePCA(df, number_of_dimension)
}
\arguments{
\item{df}{\code{CPTAC.LUAD.Proteomics.Sample.subset} generated inside
\code{\link[=.plot_PCA_CPTAC_LUAD]{.plot_PCA_CPTAC_LUAD()}}}

\item{number_of_dimension}{maximum number of components kept in the final results}
}
\value{
imputed PCA results
}
\description{
A helper function performs impute PCA on data generated by \code{\link[=.get_df_subset]{.get_df_subset()}}.
It should not be used directly, only inside \code{\link[=.plot_PCA_CPTAC_LUAD]{.plot_PCA_CPTAC_LUAD()}}
function.
}
\examples{
\dontrun{
.protein_imputePCA(CPTAC.LUAD.Proteomics.Sample.subset)
}

}
\seealso{
Other helper function for PCA: 
\code{\link{.RNAseq_PCA}()}
}
\concept{helper function for PCA}
\keyword{internal}
