% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteomics.R
\name{.protein_scatterplot}
\alias{.protein_scatterplot}
\title{Scatter plots of protein coexpression}
\usage{
.protein_scatterplot(df, protein01, protein02, color)
}
\arguments{
\item{df}{a subset LUAD dataset generated from
\code{\link[=.plot_scatterplot_protein_CPTAC]{.plot_scatterplot_protein_CPTAC()}}}

\item{protein01}{protein name}

\item{protein02}{protein name}

\item{color}{color scheme for the scatter plot}
}
\description{
A helper function for protein coexpression analysis
}
\details{
This function should not be used directly, only inside
\code{\link[=.plot_scatterplot_protein_CPTAC]{.plot_scatterplot_protein_CPTAC()}} function.

Side effects:
(1) scatter plot to show correlation between two protein expressions
}
\seealso{
Other helper function for protein coexpression analysis: 
\code{\link{.get_CPTAC_LUAD_Phosproteomics_subset}()},
\code{\link{.get_CPTAC_LUAD_Proteomics_subset}()},
\code{\link{.protein_boxplot}()}
}
\concept{helper function for protein coexpression analysis}
\keyword{internal}
