% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAProCorr.R
\name{.RNApro_scatterplot}
\alias{.RNApro_scatterplot}
\title{Scatter plots of correlation between RNAseq and proteomics data}
\usage{
.RNApro_scatterplot(df, gene_name, cohort)
}
\arguments{
\item{df}{dataframe of both RNAseq and proteomics generated inside}

\item{gene_name}{protein or gene name}

\item{cohort}{database name. \code{CCLE} or \code{LUAD}}
}
\description{
A helper function plots RNAseq and proteomics data in scatter
plot and analyzes the correlation
}
\details{
This function should not be used directly, only inside
\code{\link[=.plot_scatter_RNApro_CCLE]{.plot_scatter_RNApro_CCLE()}} or \code{\link[=.plot_scatter_RNApro_LUAD]{.plot_scatter_RNApro_LUAD()}} function.

Side effects:

(1) scatter plots to show correlation between RNA and protein
expressions
}
\seealso{
Other helper function for RNA protein correlation analysis: 
\code{\link{.get_CCLE_Proteomics_subset}()},
\code{\link{.get_CCLE_RNAseq_subset}()}
}
\concept{helper function for RNA protein correlation analysis}
\keyword{internal}
