% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{.RNAratio_calculation}
\alias{.RNAratio_calculation}
\title{Calculate RNA ratios in tumors vs adjacent normal tissues}
\usage{
.RNAratio_calculation(
  EIF4E,
  EIF4E2,
  EIF4E3,
  EIF4EBP1,
  EIF4G1,
  EIF4G2,
  EIF4G3,
  EIF3D,
  EIF4A1,
  EIF4A2
)
}
\arguments{
\item{EIF4E}{gene name, same input from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}

\item{EIF4E2}{gene name, same input from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}

\item{EIF4E3}{gene name, same input from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}

\item{EIF4EBP1}{gene name, same input from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}

\item{EIF4G1}{gene name, same input from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}

\item{EIF4G2}{gene name, same input from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}

\item{EIF4G3}{gene name, same input from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}

\item{EIF3D}{gene name, same input from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}

\item{EIF4A1}{gene name, same input from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}

\item{EIF4A2}{gene name, same input from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}
}
\value{
a data frame of differential RNA ratios in tumors vs adjacent normal tissues
from individual TCGA cancer types
}
\description{
This function calculates the RNA ratio data from TCGA samples including all
tumors and solid tissue normal samples for comparison.
It should not be used directly, only inside \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}
function.
}
\keyword{internal}
