% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{.RNAratio_tumortype}
\alias{.RNAratio_tumortype}
\title{Analyzes differential RNA ratios in primary, metastatic tumors vs
adjacent normal tissues from all TCGA cancer types combined}
\usage{
.RNAratio_tumortype(df, gene_ratio)
}
\arguments{
\item{df}{\code{.RNAratio_tumortype(.RNAratio_data)} generated
inside \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}}

\item{gene_ratio}{}
}
\value{
a data frame of differential RNA ratios in tumors vs adjacent normal tissues
from individual TCGA cancer types
}
\description{
This function selects the RNA ratio data from TCGA samples including
tumor samples that are labeled as \code{metastatic}, \verb{primary tumor},
and \verb{solid tissue normal} for comparison.

It should not be used directly, only
inside \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}} function.
}
\seealso{
Other helper function for differential expression analysis: 
\code{\link{.RNAratio_calculation}()},
\code{\link{.RNAratio_selection}()},
\code{\link{.RNAseq_all_gene}()},
\code{\link{.RNAseq_ind_gene}()},
\code{\link{.RNAseq_tumortype}()}
}
\concept{helper function for differential expression analysis}
\keyword{internal}
