% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DEG.R
\name{.RNAseq_all_gene}
\alias{.RNAseq_all_gene}
\title{Compares expressions among genes in tumor samples}
\usage{
.RNAseq_all_gene(df)
}
\arguments{
\item{df}{\code{.TCGA_GTEX_RNAseq_sampletype_subset} generated inside \code{\link{.plot_boxgraph_RNAseq_TCGA}}}
}
\value{
a data frame ranking genes by their mRNA expressions in lung adenocarcinoma
}
\description{
This function selects the RNAseq data from TCGA samples,
excludes “Solid Tissue Normal” and ranks the genes by their mRNA expression.

It should not be used directly, only inside \code{\link{.plot_boxgraph_RNAseq_TCGA}} function.
}
\examples{
\dontrun{
.RNAseq_all_gene(.TCGA_GTEX_RNAseq_sampletype_subset)
}
}
\keyword{internal}
