% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{.RNAseq_grouped_boxplot}
\alias{.RNAseq_grouped_boxplot}
\title{Grouped box plots of RNA expression across tumors}
\usage{
.RNAseq_grouped_boxplot(df)
}
\arguments{
\item{df}{output dataset from \code{.RNAseq_all_gene(.TCGA_GTEX_RNAseq_sampletype_subset)},
which was generated inside \code{\link[=.plot_boxgraph_RNAseq_TCGA]{.plot_boxgraph_RNAseq_TCGA()}}function.}
}
\description{
This function takes the output of \code{\link[=.RNAseq_all_gene]{.RNAseq_all_gene()}} function and generate
box plots of differential gene expression across all TCGA cancer types
This function should not be used directly,
only inside \code{\link[=.plot_boxgraph_RNAseq_TCGA]{.plot_boxgraph_RNAseq_TCGA()}}function.

Side effect: box plots on screen and as pdf file to compare relative
expression of EIF genes across all TCGA cancer type
}
\seealso{
Other helper function for differential expression plotting: 
\code{\link{.RNAratio_boxplot}()},
\code{\link{.RNAseq_boxplot}()},
\code{\link{.violinplot}()}
}
\concept{helper function for differential expression plotting}
\keyword{internal}
