% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.RNAseq_PCA}
\alias{.RNAseq_PCA}
\title{\subsection{Perform PCA or imputed PCA and plot the PCA results}{

Perform PCA on RNAseq data
}}
\usage{
.RNAseq_PCA(df, x)
}
\arguments{
\item{df}{\code{df} generated inside \code{\link[=.get_df_subset]{.get_df_subset()}}}

\item{x}{number of components kept in the final results}
}
\value{
PCA results
}
\description{
This function perform PCA on data generated by \code{\link[=.get_df_subset]{.get_df_subset()}}.
It should not be used directly, only inside \code{\link[=.plot_PCA_TCGA_GTEX]{.plot_PCA_TCGA_GTEX()}} or
\code{\link[=.plot_PCA_TCGA_GTEX_tumor]{.plot_PCA_TCGA_GTEX_tumor()}} function.
}
\examples{
\dontrun{
.RNAseq_PCA(df[[1]], 10)
}

}
\keyword{internal}
