% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{.RNAseq_tumortype}
\alias{.RNAseq_tumortype}
\title{Select gene expression data in primary, metastatic tumors vs adjacent normal
tissues from all TCGA cancer types combined}
\usage{
.RNAseq_tumortype(df)
}
\arguments{
\item{df}{\code{.TCGA_GTEX_RNAseq_sampletype_subset} generated inside
\code{\link[=.plot_boxgraph_RNAseq_TCGA]{.plot_boxgraph_RNAseq_TCGA()}}}
}
\value{
a data frame of differential gene expression in tumors vs adjacent
normal tissues in individual TCGA cancer types
}
\description{
This function selects the RNAseq data from TCGA samples including
tumor samples that are labeled as \code{metastatic}, \verb{primary tumor},
and \verb{solid tissue normal} for comparison.
It should not be used directly, only inside \code{\link[=.plot_boxgraph_RNAseq_TCGA]{.plot_boxgraph_RNAseq_TCGA()}}
function.
}
\seealso{
Other helper function for differential expression analysis: 
\code{\link{.RNAratio_calculation}()},
\code{\link{.RNAratio_selection}()},
\code{\link{.RNAratio_tumortype}()},
\code{\link{.RNAseq_all_gene}()},
\code{\link{.RNAseq_ind_gene}()}
}
\concept{helper function for differential expression analysis}
\keyword{internal}
