% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.summarize_counts}
\alias{.summarize_counts}
\title{Summary of total number of CORs identified for each EIF4F gene}
\usage{
.summarize_counts(identified_COR, COR_type)
}
\arguments{
\item{identified_COR}{a data frame with logic statement for each gene tested}

\item{COR_type}{posCORs or negCORs}
}
\value{
a table with counts of posCORs or negCORs for each EIF4F gene
}
\description{
This function summarizes the total number of posCORs or negCORs identified for
each EIF4F gene

It should not be used directly, only inside \code{\link[=.EIF_correlation]{.EIF_correlation()}}
function.
}
\seealso{
Other helper function to identify correlating genes for EIF4F genes: 
\code{\link{.EIF_correlation}()},
\code{\link{.correlation_coefficient}()},
\code{\link{.is_significant_negcor}()},
\code{\link{.is_significant_poscor}()}
}
\concept{helper function to identify correlating genes for EIF4F genes}
\keyword{internal}
