% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{EIF4F_CNV_analysis}
\alias{EIF4F_CNV_analysis}
\title{Perform all CNV related analysis and generate plots}
\usage{
EIF4F_CNV_analysis()
}
\value{
CNV analysis plots
}
\description{
A wrapper function to call all master functions for CNV analysis with inputs
}
\details{
This function run three master functions together:
\code{\link{.plot_bargraph_CNV_TCGA}},
\code{\link{.plot_matrix_CNVcorr_TCGA}} and
\code{\link{.plot_boxgraph_CNVratio_TCGA}} with inputs
}
\examples{
\dontrun{
EIF4F_CNV_analysis()
}
}
