% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{EIF4F_DEG_analysis}
\alias{EIF4F_DEG_analysis}
\title{Perform all CNV related analysis and generate plots}
\usage{
EIF4F_DEG_analysis()
}
\value{
DEG analysis plots
}
\description{
A wrapper function to call all master functions for CNV analysis with inputs
}
\details{
This function run two master functions together, with inputs:
\itemize{
\item \code{\link[=.plot_boxgraph_RNAseq_TCGA]{.plot_boxgraph_RNAseq_TCGA()}}
\item \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}
}
}
\examples{
\dontrun{
EIF4F_DEG_analysis()
}

}
