% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteincorr.R
\name{EIF4F_Proteomics_analysis}
\alias{EIF4F_Proteomics_analysis}
\title{\subsection{Wrapper function to call all master functions with inputs}{

Analyze co-expression among EIF4F subunits and differential expression
}}
\usage{
EIF4F_Proteomics_analysis()
}
\value{
proteomics data analysis plots
}
\description{
A wrapper function to call all master functions for protein coexpression and
differential expression.
}
\details{
This function run the master functions \code{\link[=.plot_scatterplot_protein_LUAD]{.plot_scatterplot_protein_LUAD()}} and
\code{\link[=.plot_boxgraph_protein_CPTAC]{.plot_boxgraph_protein_CPTAC()}} with EIF4F gene name as inputs.
}
\examples{
\dontrun{
EIF4F_Proteomics_analysis()
}

}
