% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAProCorr.R
\name{EIF4F_RNA_pro_correlation}
\alias{EIF4F_RNA_pro_correlation}
\title{Perform RNA protein correlation and generate scatter plots}
\usage{
EIF4F_RNA_pro_correlation()
}
\description{
A wrapper function to call all composite functions for RNA and protein
correlation with inputs.
}
\details{
This function run the composite functions \code{\link[=.plot_scatter_RNApro_CCLE]{.plot_scatter_RNApro_CCLE()}} and
\code{\link[=.plot_scatter_RNApro_LUAD]{.plot_scatter_RNApro_LUAD()}} with EIF4F gene name as inputs.

Side effects:

(1) scatter plot to show correlation between EIF4F RNA and protein
expressions in CCLE samples

(2) scatter plot to show correlation between EIF4F RNA and protein
expressions in CPTAC LUAD samples
}
\examples{
\dontrun{
EIF4F_RNA_pro_correlation()
}

}
\seealso{
Other wrapper function to call all composite functions with inputs: 
\code{\link{EIF4F_CNV_analysis}()},
\code{\link{EIF4F_Corrgene_analysis}()},
\code{\link{EIF4F_DEG_analysis}()},
\code{\link{EIF4F_PCA}()},
\code{\link{EIF4F_Proteomics_analysis}()},
\code{\link{EIF4F_Survival_analysis}()}
}
\concept{wrapper function to call all composite functions with inputs}
