% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{initialize_cnv_data}
\alias{initialize_cnv_data}
\title{Read all CNV related datasets from TCGA}
\usage{
initialize_cnv_data()
}
\description{
This function reads all CNV related datasets from TCGA and generates three global variables.

TCGA_CNV_value: the unthreshold CNV value data generated from \code{\link{.get_TCGA_CNV_value}}

TCGA_CNV_sampletype: the merged dataset from .TCGA.CNV, the threshold CNV data generated from \code{\link{.get_TCGA_CNV}},
and .TCGA_sampletype, the annotation data from the "TCGA_phenotype_denseDataOnlyDownload.tsv"
dataset with selection of sample.type and primary.disease columns. “Solid Tissue Normal” samples are excluded.

TCGA_CNVratio_sampletype: the merged dataset from .TCGA_CNV_ratio, the CNV ratio data generated from \code{\link{.get_TCGA_CNV_ratio}},
and .TCGA_sampletype.
}
\examples{
\dontrun{
initialize_cnv_data()
}

}
