% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{initialize_dir}
\alias{initialize_dir}
\title{Set output directories}
\usage{
initialize_dir()
}
\description{
A wrapper function sets up the output directories
}
\details{
Side effects:

(1) output directory for CNV analysis results \verb{~/Documents/EIF_output/CNV}

(2) output directory for differential gene expression and ratio analysis
results \verb{~/Documents/EIF_output/DEG}

(3) output directories for survival analysis results
\itemize{
\item \verb{~/Documents/EIF_output/Survival/KM}
\item \verb{~/Documents/EIF_output/Survival/CoxPH}
}

(4) output directory for PCA results
\itemize{
\item \verb{~/Documents/EIF_output/PCA/All}
\item \verb{~/Documents/EIF_output/PCA/TCGA}
\item \verb{~/Documents/EIF_output/PCA/GTEX}
\item \verb{~/Documents/EIF_output/PCA/Lung}
}

(5) output directory for analysis results on EIF4F correlating genes
\verb{~/Documents/EIF_output/CORs}

(6) output directory for correlation analysis between RNA and protein levels
\verb{~/Documents/EIF_output/RNApro}

(7) output directory for co-expression analysis among EIF4F subunits and
their differential expression \verb{~/Documents/EIF_output/Proteomics}
}
\examples{
\dontrun{
initialize_dir()
}

}
\concept{wrapper function for setting up output directories}
