% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{initialize_format}
\alias{initialize_format}
\title{Set format for font and color of plots}
\usage{
initialize_format()
}
\description{
A wrapper function set up the font type, size and color for
ggplots.
}
\examples{
\dontrun{
initialize_format()
}

}
\concept{wrapper function for format initialization}
