% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteomics.R
\name{initialize_phosphoproteomics_data}
\alias{initialize_phosphoproteomics_data}
\title{Read all phosphoproteomics related datasets from CPTAC LUAD}
\usage{
initialize_phosphoproteomics_data()
}
\description{
A wrapper function reads all phosphoproteomics related datasets
from CPTAC LUAD.
}
\details{
Side effects:

(1) \code{CPTAC_LUAD_Phos}: the phosphoproteomics data of CPTAC LUAD from the
download data \code{Phos.xlsx}.

(2) \code{CPTAC_LUAD_Clinic_Sampletype}: a merged dataset from two data frames.
It provides the annotation data about sample types (tumor or healthy tissues)
and tumors stages of each sample
\itemize{
\item \code{.CPTAC_LUAD_Clinic}, the CPTAC clinical data from the download dataset
\code{S046_BI_CPTAC3_LUAD_Discovery_Cohort_Clinical_Data_r1_May2019.xlsx}
\item \code{.CPTAC_LUAD_sampletype}, the CPTAC annotation data from the download
dataset \code{S046_BI_CPTAC3_LUAD_Discovery_Cohort_Samples_r1_May2019.xlsx}
}

\code{CPTAC_LUAD_Phos} and \code{CPTAC_LUAD_Clinic_Sampletype} are
stored as \code{CPTAC_LUAD_Phos.csv} and \code{CPTAC_LUAD_Clinic_Sampletype.csv} in
\verb{~/Documents/EIF_output/ProcessedData} folder.
}
\examples{
\dontrun{
initialize_phosphoproteomics_data()
}

}
\seealso{
Other wrapper function for data initialization: 
\code{\link{initialize_RNAseq_data}()},
\code{\link{initialize_cnv_data}()},
\code{\link{initialize_data}()},
\code{\link{initialize_proteomics_data}()},
\code{\link{initialize_survival_data}()}
}
\concept{wrapper function for data initialization}
