% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{initialize_RNAseq_data}
\alias{initialize_RNAseq_data}
\title{Read all RNA-seq related datasets from TCGA and GTEx}
\usage{
initialize_RNAseq_data()
}
\description{
A wrapper function reads RNA-seq related datasets from TCGA and GTEx.
}
\details{
Its side effects is the global variable \code{TCGA_GTEX_RNAseq_sampletype}, which
was merged from two internal data frames:

(1) \code{.TCGA_GTEX_RNAseq}: the recomputed RNAseq data from both TCGA and GTEx
generated by \code{\link[=.get_TCGA_GTEX_RNAseq]{.get_TCGA_GTEX_RNAseq()}}, which imports the dataset
\code{TcgaTargetGtex_RSEM_Hugo_norm_count}.

(2) \code{.TCGA_GTEX_sampletype} annotates the feature for each sample from
TCGA and GTEx. The data frame imports the \code{TcgaTargetGTEX_phenotype.txt}
dataset and performed basic data cleaning steps including removal of
duplicates and NAs.

To reduce the data size, we only select the following four relevant columns
out of \code{TcgaTargetGTEX_phenotype.txt} to construct \code{.TCGA_GTEX_sampletype}.
\itemize{
\item \code{sample.type} column that annotates  malignant of normal tissues
\item \code{primary.disease} column that annotates cancer types for each sample
\item \code{primary.site} column that annotates the tissue types
\item \code{study} column that annotates the cohort “TCGA” or “GTEx”
}

\code{TCGA_GTEX_RNAseq_sampletype} was stored as \code{TCGA_GTEX_RNAseq_sampletype.csv} in
\verb{~/Documents/EIF_output/ProcessedData} folder.
}
\examples{
\dontrun{
initialize_RNAseq_data()
}

}
\seealso{
Other wrapper function for data initialization: 
\code{\link{initialize_cnv_data}()},
\code{\link{initialize_data}()},
\code{\link{initialize_phosphoproteomics_data}()},
\code{\link{initialize_proteomics_data}()},
\code{\link{initialize_survival_data}()}
}
\concept{wrapper function for data initialization}
