% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-Survival.R
\name{initialize_survival_data}
\alias{initialize_survival_data}
\title{Read RNA-seq and survival datasets from TCGA}
\usage{
initialize_survival_data()
}
\description{
This function reads RNA-seq and survival datasets from TCGA

.TCGA_RNAseq: the RNAseq data from TCGA generated from \code{\link{.get_TCGA_RNAseq}}

.TCGA_OS: the annotation data from the "Survival_SupplementalTable_S1_20171025_xena_sp"
dataset with selection of OS and OS.time columns.

.TCGA_sampletype: the annotation data from the "TCGA_phenotype_denseDataOnlyDownload.tsv"
dataset with selection of sample.type and primary.disease columns.

TCGA_RNAseq_OS_sampletype: the merged dataset from .TCGA_RNAseq, .TCGA_OS and .TCGA_sampletype
}
\examples{
\dontrun{
initialize_survival_data()
}

}
