% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{plot.bargraph.CNV.TCGA}
\alias{plot.bargraph.CNV.TCGA}
\title{Summary of CNV statuses in bar plots}
\usage{
plot.bargraph.CNV.TCGA(EIF)
}
\arguments{
\item{EIF}{gene names in a vector of characters}
}
\value{
stacked bar plots for grouped CNV status of \code{EIF} in TCGA tumors
}
\description{
Provides the summary of CNV statuses of EIF4F genes in tumors from all TCGA
cancer types combined and in tumors from individual TCGA cancer types
}
\details{
This function first selects CNV and sample type data from only EIF4F gene
in the data frame \code{TCGA.CNV.sampletype} prepared from \code{\link{initialize.cnv.data}}.

Then it uses the subset data \code{TCGA.CNV.sampletype.subset} to perform the CNV status analysis of all inquired genes
with the functions \code{\link{.CNV.all.cancer}} and plot the results as a bar plot with \code{\link{.CNV.sum.barplot}}

It also uses the same subset data to perform CNV status analysis for each gene across
all tumors types with the function \code{\link{.CNV.ind.cancer}} and plots the CNV results and \code{\link{.CNV.barplot}}
}
\examples{
\dontrun{plot.bargraph.CNV.TCGA(c("EIF4A1", "EIF4E", "EIF4EBP1", "EIF4G1"))}
}
