% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{plot.boxgraph.CNVratio.TCGA}
\alias{plot.boxgraph.CNVratio.TCGA}
\title{CNV ratios in tumors vs adjacent normal tissues across tumor types}
\usage{
plot.boxgraph.CNVratio.TCGA(EIF)
}
\arguments{
\item{EIF}{gene names in a vector of characters}
}
\value{
stacked bar plots for grouped CNV status of \code{EIF} in TCGA tumors
}
\description{
This function generates boxplot for CNV ratios in tumors vs adjacent normals from individual TCGA cancer types.
}
\details{
This function first selects CNV and sample type data from only EIF4F gene
in the data frame \code{TCGA.CNVratio.sampletype} prepared from \code{\link{initialize.cnv.data}}.

Then it uses the subset data \code{TCGA.CNVratio.sampletype.subset} to perform CNV status analysis for each gene across
all tumors types with the function \code{\link{.CNVratio.tumor}} and plots the CNV results and \code{\link{.CNVratio.boxplot}}
}
\examples{
\dontrun{plot.boxgraph.CNVratio.TCGA(c("EIF4A1", "EIF4E", "EIF4EBP1", "EIF4G1"))}
}
