% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{plot.matrix.CNVcorr.TCGA}
\alias{plot.matrix.CNVcorr.TCGA}
\title{Correlation matrix for CNV values}
\usage{
plot.matrix.CNVcorr.TCGA(EIF)
}
\arguments{
\item{EIF}{gene names in a vector of characters}
}
\value{
the correlation matrix plot for \code{EIF} CNV values in TCGA tumors
}
\description{
generates correlation matrix for eIF4F CNV in tumors from all TCGA cancer type combined
}
\details{
This function first selects unthreshold CNV values from only EIF4F gene
in the data frame \code{TCGA.CNV.value} prepared from \code{\link{initialize.cnv.data}}.

Then it uses the subset data to calculate the correlation coefficients
between every two genes and plot the correlation matrix with the function \code{\link{.matrix.plot}}
}
\examples{
\dontrun{plot.matrix.CNVcorr.TCGA(c("EIF4A1", "EIF4E", "EIF4EBP1", "EIF4G1"))}
}
