% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{color}
\alias{color}
\title{Generates a number of most distinctive colors.}
\usage{
color()
}
\value{
a vector with 74 different colors
}
\description{
A helper function generates a number of most distinctive colors
for plotting
}
\details{
This function should not be used directly,
only inside \code{\link[=initialize_format]{initialize_format()}} function.
}
\examples{
\dontrun{
color()
}

}
\keyword{internal}
