% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.biplot}
\alias{.biplot}
\title{Plot PCA results as biplot, scree and matrix plots}
\usage{
.biplot(res.pca, df, x, y, color, folder)
}
\arguments{
\item{res.pca}{PCA results from selected RNAseq data}

\item{df}{selected RNAseq data with sample type annotation}

\item{x}{selected sample types for plot label}

\item{y}{type of primary diseases or healthy tissues for color of individuals}

\item{color}{color scheme used for individual sample on the PCA}

\item{folder}{sub directory name to store the output files}
}
\value{
PCA bioplot, matrix, scree plot
}
\description{
This function draw biplot, screen and matrix plots for PCA results
from selected RNAseq data generated from \code{\link[=.get_df_subset]{.get_df_subset()}}.
}
\examples{
\dontrun{
.biplot(
  res.pca = .RNAseq_PCA(df[[1]], 10), df = df[[2]],
  x = "Metastatic Tumor (TCGA)",
  y = "primary.disease", color = col_vector, folder = "TCGA"
)
}

}
\keyword{internal}
