% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.CNVratio_boxplot}
\alias{.CNVratio_boxplot}
\title{Box plots of the CNV ratios in tumors vs adjacent normals}
\usage{
.CNVratio_boxplot(df)
}
\arguments{
\item{df}{\code{.EIF_CNVratio_ind_cancer} generated
inside \code{\link[=.plot_boxgraph_CNVratio_TCGA]{.plot_boxgraph_CNVratio_TCGA()}}}
}
\value{
boxplot for CNV ratios in tumors vs adjacent normals
from individual TCGA cancer types.
}
\description{
This function generates boxplot for CNV ratios in tumors
vs adjacent normals from individual TCGA cancer types.
}
\details{
This plot function uses dataset \code{.TCGA_CNVratio_sampletype_subset}
generated from \code{\link[=.plot_boxgraph_CNVratio_TCGA]{.plot_boxgraph_CNVratio_TCGA()}} function.
It should not be used directly, only inside \code{\link[=.plot_boxgraph_CNVratio_TCGA]{.plot_boxgraph_CNVratio_TCGA()}} function.
}
\examples{
\dontrun{
lapply(.EIF_CNVratio_ind_cancer, .CNVratio_boxplot)
}

}
\keyword{internal}
