% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.CORs_vennplot}
\alias{.CORs_vennplot}
\title{Plot Venn diagrams for correlating genes}
\usage{
.CORs_vennplot(df, tissue_type, sample_type, CORs_type)
}
\arguments{
\item{df}{correlation data}

\item{tissue_type}{input argument of \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}}

\item{sample_type}{tumor or normal for the title of Venn diagram}

\item{CORs_type}{posCOR or negCORs for the title of Venn diagram}
}
\description{
This function draw Venn diagrams, using the correlation data generated from
\code{\link[=.EIF_correlation]{.EIF_correlation()}}.

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}
function.

Side effects:

(1) Venn diagrams on screen and as pdf file to show the overlap of
EIF correlating genes
}
\examples{
\dontrun{
.CORs_vennplot(
df = EIF.cor.tumor[[3]],
tissue_type = tissue_type,
sample_type = "tumor",
CORs_type = "posCOR")
}

}
\seealso{
Other helper function for correlation analysis plotting: 
\code{\link{.CORs_coeff_heatmap}()},
\code{\link{.CORs_summary_bargraph}()},
\code{\link{.pathway_dotplot}()}
}
\concept{helper function for correlation analysis plotting}
\keyword{internal}
