% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.EIF_correlation}
\alias{.EIF_correlation}
\title{Identify EIF4F correlating genes}
\usage{
.EIF_correlation(df, y)
}
\arguments{
\item{df}{the dataframe \code{.TCGA_GTEX_RNAseq_sampletype_subset} generated inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}}

\item{y}{sample types, either \code{all.tumor.type} or \code{c("Normal Tissue")}
generated inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}}}
}
\value{
a list output with four elements: cor.data for the heatmap function and CORs.counts for
bargraph function, c4.posCOR, c4.negCOR for Venn plots
}
\description{
This function calculate the correlation efficiency between each EIF4F gene and the rest of cellular mRNAs.
and identify the significantly correlating genes.

It should not be used directly, only inside \code{\link[=.plot_Corr_RNAseq_TCGA_GTEX]{.plot_Corr_RNAseq_TCGA_GTEX()}} function.
}
\examples{
\dontrun{
.EIF_correlation(
  df = .TCGA_GTEX_RNAseq_sampletype_subset,
  y = all.tumor.type
)
}
}
\keyword{internal}
