% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProRNACorrelation.R
\name{.get_CCLE_Proteomics_subset}
\alias{.get_CCLE_Proteomics_subset}
\title{Select subset of CCLE proteomics data}
\usage{
.get_CCLE_Proteomics_subset(EIF)
}
\arguments{
\item{EIF}{protein name}
}
\value{
a data frame of CCLE proteomics data from the input \code{EIF} genes
}
\description{
This function selected the CCLE proteomics data from the input \code{EIF}.
}
\details{
The function should not be used directly, only inside
\code{\link[=.plot_scatter_RNApro_CCLE]{.plot_scatter_RNApro_CCLE()}} function.
}
\examples{
\dontrun{
.get_CCLE_Proteomics_subset()
}

}
\keyword{internal}
