% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-ProRNACorrelation.R
\name{.get_CCLE_RNAseq_subset}
\alias{.get_CCLE_RNAseq_subset}
\title{Select subset of CCLE RNAseq data}
\usage{
.get_CCLE_RNAseq_subset(EIF)
}
\arguments{
\item{EIF}{gene name}
}
\value{
a data frame of CCLE RNAseq data from the input \code{EIF} genes
}
\description{
This function selected the CCLE RNAseq data from the input \code{EIF}.
}
\details{
The function should not be used directly, only inside \code{\link{.plot_scatter_RNApro_CCLE}} function.
}
\examples{
\dontrun{
.get_CCLE_RNAseq_subset()
}
}
\keyword{internal}
