% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-PCA.R
\name{.get_df_subset}
\alias{.get_df_subset}
\title{Select RNAseq data based on the sample types}
\usage{
.get_df_subset(df, x)
}
\arguments{
\item{df}{\code{.TCGA_GTEX_sampletype_subset} generated inside \code{\link{.plot_PCA_TCGA_GTEX}}}

\item{x}{selected sample types}
}
\value{
a subset data frame from .TCGA_GTEX_sampletype_subset for indicate sample types
}
\description{
This function selects the RNAseq data from different sample types:
primary tumors (TCGA), metastatic tumors (TCGA), healthy tissues (GTEX)
or all sample types combined (all), from dataframe \code{.TCGA_GTEX_sampletype_subset}.

It should not be used directly, only inside \code{\link{.plot_PCA_TCGA_GTEX}} or
\code{\link{.plot_PCA_TCGA_GTEX_tumor}} function.
}
\examples{
\dontrun{
.get_df_subset(.TCGA_GTEX_sampletype_subset, "Primary Tumor (TCGA)")
}
}
\keyword{internal}
