% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.get_df_subset}
\alias{.get_df_subset}
\title{Select RNAseq data based on the sample types}
\usage{
.get_df_subset(df, sample_type)
}
\arguments{
\item{df}{\code{.TCGA_GTEX_sampletype_subset} generated inside
\code{\link[=.plot_PCA_TCGA_GTEX]{.plot_PCA_TCGA_GTEX()}}}

\item{sample.type}{selected sample types}
}
\value{
a subset dataframe from \code{.TCGA_GTEX_sampletype_subset} for indicate
sample types
}
\description{
A helper function selects the RNAseq data of primary tumors (TCGA), metastatic
tumors (TCGA), healthy tissues (GTEX) or all sample types combined (all),
from dataframe \code{.TCGA_GTEX_sampletype_subset}.

It should not be used directly, only inside \code{\link[=.plot_PCA_TCGA_GTEX]{.plot_PCA_TCGA_GTEX()}} or
\code{\link[=.plot_PCA_TCGA_GTEX_tumor]{.plot_PCA_TCGA_GTEX_tumor()}} function.
}
\examples{
\dontrun{
.get_df_subset(.TCGA_GTEX_sampletype_subset, "Primary Tumor (TCGA)")
}

}
\concept{helper function to select RNA-seq data from specific study groups
for PCA}
\keyword{internal}
