% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DEG.R
\name{.get_TCGA_GTEX_RNAseq}
\alias{.get_TCGA_GTEX_RNAseq}
\title{Read recomputed RNAseq data from both TCGA and GTEx}
\usage{
.get_TCGA_GTEX_RNAseq()
}
\value{
a data frame that contains the recomputed RNAseq data from both TCGA and GTEx
}
\description{
This function reads recomputed RNAseq data from both TCGA and GTEx
"TcgaTargetGtex_RSEM_Hugo_norm_count".
}
\details{
The function also removes possible duplicated tumor samples and samples with NAs in the dataset.

It should not be used directly, only inside \code{\link{initialize_RNAseq_data}} function.
}
\examples{
\dontrun{
.get_TCGA_GTEX_RNAseq()
}
}
\keyword{internal}
