% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-Survival.R
\name{.get_TCGA_RNAseq}
\alias{.get_TCGA_RNAseq}
\title{Read the RNAseq data from TCGA}
\usage{
.get_TCGA_RNAseq()
}
\value{
a data frame that contains the RNAseq data from TCGA
}
\description{
This function reads the RNAseq data from TCGA
"EB++AdjustPANCAN_IlluminaHiSeq_RNASeqV2.geneExp.xena".
}
\details{
The function also removes possible duplicated tumor samples and samples with NAs in the dataset.

It should not be used directly, only inside \code{\link{initialize_survival_data}} function.
}
\examples{
\dontrun{
.get_TCGA_RNAseq()
}
}
\keyword{internal}
