% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{.KM_curve}
\alias{.KM_curve}
\title{\subsection{Survival analysis and plotting}{

Kaplan Meier survival analyses of gene expression
}}
\usage{
.KM_curve(gene, data, cutoff, tumor)
}
\arguments{
\item{gene}{gene name, passed \code{EIF} argument from \code{\link[=.plot_KM_RNAseq_TCGA]{.plot_KM_RNAseq_TCGA()}}}

\item{data}{\code{df} generated inside \code{\link[=.plot_KM_RNAseq_TCGA]{.plot_KM_RNAseq_TCGA()}}}

\item{cutoff}{percentage of gene expression}

\item{tumor}{all tumor types or specific type}
}
\value{
a KM plot
}
\description{
This function correlate the gene expression within tumor samples with patient
overall survival time from TCGA tumors.
}
\details{
It should not be used directly, only inside \code{\link[=.plot_KM_RNAseq_TCGA]{.plot_KM_RNAseq_TCGA()}}
function.
}
\examples{
\dontrun{
.KM_curve(gene = EIF, data = df, cutoff = cutoff, tumor = tumor)
}

}
\keyword{internal}
