% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{.matrix_plot}
\alias{.matrix_plot}
\title{Correlation coefficients for CNV values}
\usage{
.matrix_plot(df)
}
\arguments{
\item{df}{output of \code{TCGA_CNV_value \%>\% select(all_of(EIF))} generated inside \code{\link{.plot_matrix_CNVcorr_TCGA}}}
}
\value{
stacked bar plots for CNV status of each gene in an individual cancer type.
}
\description{
This function calculates the correlation coefficients between every two genes and plot the correlation matrix with the function.
}
\details{
This plot function uses dataset \code{TCGA_CNV_value} generated from \code{\link{initialize_cnv_data}} function
It should not be used directly, only inside \code{\link{.plot_matrix_CNVcorr_TCGA}} function.
}
\examples{
\dontrun{
TCGA_CNV_value \%>\%
  dplyr::select(all_of(EIF_list)) \%>\%
  .matrix_plot()
}
}
\keyword{internal}
