% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.plot_boxgraph_CNVratio_TCGA}
\alias{.plot_boxgraph_CNVratio_TCGA}
\title{CNV ratios in tumors vs adjacent normal tissues across tumor types}
\usage{
.plot_boxgraph_CNVratio_TCGA(gene_list)
}
\arguments{
\item{gene_list}{gene names in a vector of characters}
}
\description{
This function generates boxplot for CNV ratios in tumors vs adjacent normals
from individual TCGA cancer types.
}
\details{
This function
\itemize{
\item selects CNV and sample type data
from only EIF4F gene in the data frame \code{TCGA_CNVratio_sampletype}
prepared from \code{\link[=initialize_cnv_data]{initialize_cnv_data()}}.
\item uses the subset data \code{.TCGA_CNVratio_sampletype_subset}
to perform CNV status analysis for each gene across
all tumors types with the function \code{\link[=.CNVratio_tumor]{.CNVratio_tumor()}}
and plots the CNV results and \code{\link[=.CNVratio_boxplot]{.CNVratio_boxplot()}}
}

It should not be used directly, only inside \code{\link[=EIF4F_CNV_analysis]{EIF4F_CNV_analysis()}} function.

Side effects:

(1) the box plots on screen and as saved pdf files to show CNV ratio of
\code{gene_list} in TCGA tumors vs normals
}
\examples{
\dontrun{
plot_boxgraph_CNVratio_TCGA(c("EIF4A1", "EIF4E", "EIF4EBP1", "EIF4G1"))
}

}
\seealso{
Other composite function to call CNV data analysis and plotting: 
\code{\link{.plot_bargraph_CNV_TCGA}()}
}
\concept{composite function to call CNV data analysis and plotting}
\keyword{internal}
