% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteincorr.R
\name{.plot_boxgraph_protein_CPTAC}
\alias{.plot_boxgraph_protein_CPTAC}
\title{Comparison of protein and phosphorylation levels among different stages of
LUAD}
\usage{
.plot_boxgraph_protein_CPTAC(EIF_list)
}
\arguments{
\item{EIF_list}{protein names in a vector of string}
}
\value{
the box plots for \code{EIF} protein and phosphorylation values in LUAD
}
\description{
This function generates boxplot for eIF4F proteomics and prosphorylation data
from CPTAC LUAD.
}
\details{
This function merges the LUAD proteomics values from EIF4F genes
in the data frame prepared from \code{\link[=.get_CPTAC_LUAD_Proteomics_subset]{.get_CPTAC_LUAD_Proteomics_subset()}} and
phosphoproteomics data of the same protein in the data frame prepared
from \code{\link[=.get_CPTAC_LUAD_Phosproteomics_subset]{.get_CPTAC_LUAD_Phosproteomics_subset()}}.

Then it uses the combined data to compare the abundance across different
tumor stages, and plot the result with the function \code{\link[=.protein_boxplot]{.protein_boxplot()}}

This function should not be used directly, only inside
\code{\link[=EIF4F_Proteomics_analysis]{EIF4F_Proteomics_analysis()}} function.
}
\examples{
\dontrun{
.plot_boxgraph_protein_CPTAC(c(
  "EIF4G1", "EIF4A1", "EIF4E", "EIF4EBP1",
  "AKT1", "MTOR", "EIF4B", "EIF4H", "MKNK1", "MKNK2"
))
}

}
\keyword{internal}
