% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DEG.R
\name{.plot_boxgraph_RNAratio_TCGA}
\alias{.plot_boxgraph_RNAratio_TCGA}
\title{Compare the RNA ratios between EIF4F genes}
\usage{
.plot_boxgraph_RNAratio_TCGA(
  EIF4E,
  EIF4E2,
  EIF4E3,
  EIF4EBP1,
  EIF4G1,
  EIF4G2,
  EIF4G3,
  EIF3D,
  EIF4A1,
  EIF4A2
)
}
\arguments{
\item{EIF4E}{gene name as a string}

\item{EIF4E2}{gene name as a string}

\item{EIF4E3}{gene name as a string}

\item{EIF4EBP1}{gene name as a string}

\item{EIF4G1}{gene name as a string}

\item{EIF4G2}{gene name as a string}

\item{EIF4G3}{gene name as a string}

\item{EIF3D}{gene name as a string}

\item{EIF4A1}{gene name as a string}

\item{EIF4A2}{gene name as a string}
}
\value{
box plots for RNA ratios among input argument in TCGA tumors
}
\description{
This function generates box plot to compare the RNA ratios between EIF4F genes in tumors tumors vs adjacent normal tissues across TCGA cancer types.
and a violin plot to compare RNA ratios in primary, metastatic tumors vs adjacent normal tissues from all TCGA cancer types combined.
}
\details{
This function analyzes the ratios of mRNA levels between input genes within each TCGA sample,
in the data frame \code{TCGA_GTEX_RNAseq_sampletype} prepared from \code{\link{initialize_RNAseq_data}}.

It calculates RNA ratio across individual tumor types from all cancers
with the function \code{\link{.RNAratio_calculation}}, select the ratio data with \code{\link{.RNAratio_selection}}
and plots with \code{\link{.RNAratio_boxplot}}.

Finally, it compares RNA ratio in primary, metastatic tumors vs adjacent normal tissues from all TCGA cancer types combined with
the function \code{\link{.RNAratio_tumortype}} and plots with \code{\link{.violinplot}}.
It should not be used directly, only inside \code{\link{EIF4F_DEG_analysis}} function.
}
\examples{
\dontrun{
.plot_boxgraph_RNAratio_TCGA(
  EIF4E = "EIF4E", EIF4E2 = "EIF4E2",
  EIF4E3 = "EIF4E3", EIF4EBP1 = "EIF4EBP1", EIF4G1 = "EIF4G1", EIF4G2 = "EIF4G2",
  EIF4G3 = "EIF4G3", EIF3D = "EIF3D", EIF4A1 = "EIF4A1", EIF4A2 = "EIF4A2"
)
}
}
