% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{.plot_boxgraph_RNAseq_TCGA}
\alias{.plot_boxgraph_RNAseq_TCGA}
\title{\subsection{master functions to call DEG gene analysis and plotting}{

Compare the expression of EIF4F genes
}}
\usage{
.plot_boxgraph_RNAseq_TCGA(EIF.list)
}
\arguments{
\item{EIF.list}{gene names in a vector of characters}
}
\value{
box plots for differential gene expression of \code{EIF.list} in TCGA tumors
}
\description{
This function generates a summary box plot to compare the expression of all
EIF4F genes across TCGA cancer types, box plots for differential gene
expression in tumors tumors vs adjacent normal tissues for each  gene.
and a violin plot to compare differential gene expression in primary,
metastatic tumors vs adjacent normal tissues from all TCGA cancer types
combined.
}
\details{
This function
\itemize{
\item selects RNASeq and sample type data from input gene
in the data frame \code{TCGA_GTEX_RNAseq_sampletype} prepared
from \code{\link[=initialize_RNAseq_data]{initialize_RNAseq_data()}}.
\item With the subset data \code{.TCGA_GTEX_RNAseq_sampletype_subset}, compares the
mRNA expression of all inquired genes with the functions \code{\link[=.RNAseq_all_gene]{.RNAseq_all_gene()}}
and plot the results as a bar plot with \code{\link[=.RNAseq_grouped_boxplot]{.RNAseq_grouped_boxplot()}}
\item performs differential expression analysis for each gene across all
tumors types with the function \code{\link[=.RNAseq_ind_gene]{.RNAseq_ind_gene()}} and plots with
\code{\link[=.RNAseq_boxplot]{.RNAseq_boxplot()}}.
\item performs differential expression analysis for each gene in
primary, metastatic tumors vs adjacent normal tissues from all TCGA cancer
types combined with the function \code{\link[=.RNAseq_tumortype]{.RNAseq_tumortype()}} and plots with
\code{\link[=.violinplot]{.violinplot()}}.
}

It should not be used directly, only inside \code{\link[=EIF4F_DEG_analysis]{EIF4F_DEG_analysis()}} function.
}
\examples{
\dontrun{
.plot_boxgraph_RNAseq_TCGA(c(
  "EIF4G1", "EIF4G2", "EIF4G3",
  "PABPC1", "EIF4A1", "EIF4A2", "EIF4B", "EIF4H", "EIF4E",
  "EIF4E2", "EIF4E3", "EIF4EBP1", "EIF3D"
))
}

}
\keyword{internal}
