% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COR.R
\name{.plot_Corr_RNAseq_TCGA_GTEX}
\alias{.plot_Corr_RNAseq_TCGA_GTEX}
\title{Perform correlation analysis on RNAseq data from all tumors and
healthy tissues}
\usage{
.plot_Corr_RNAseq_TCGA_GTEX(tissue_type)
}
\arguments{
\item{tissue_type}{all tissue/tumor types or one specific type}
}
\description{
This function
\itemize{
\item finds correlating genes of EIF4F from the data frame \code{TCGA_GTEX_RNAseq_sampletype}
}
\itemize{
\item a global variable generated by \code{\link[=initialize_RNAseq_data]{initialize_RNAseq_data()}}.
}
\itemize{
\item identifies positively or negatively correlating genes from tumor or
healthy samples by \code{\link[=.EIF_correlation]{.EIF_correlation()}},
\item finds the overlap of CORs for EIF4F subunits with \code{\link[=.CORs_vennplot]{.CORs_vennplot()}},
\item compares the number of CORs for EIF4F subunits with
\code{\link[=.combine_CORs_summary]{.combine_CORs_summary()}} and plot the results in bargraph
with \code{\link[=.CORs_summary_bargraph]{.CORs_summary_bargraph()}}
\item compares the correlation strength of CORs from healthy tissues and tumors
from merged data frame generated from \code{\link[=.combine_CORs_list]{.combine_CORs_list()}} and visualizes
clustering results in heatmap with \code{\link[=.CORs_coeff_heatmap]{.CORs_coeff_heatmap()}}
\item retrieve the gene names from the clusters in heatmap with
\code{\link[=.get_cluster_genes]{.get_cluster_genes()}}
\item performs the pathways enrichment analysis on the retrieved gene list and
plot the results with \code{\link[=.pathway_dotplot]{.pathway_dotplot()}}
}

This function is not accessible to the user and will not show at the users'
workspace. It can only be called by the exported \code{\link[=EIF4F_Corrgene_analysis]{EIF4F_Corrgene_analysis()}}
function.

Side effects:

(1) Venn diagrams on screen and as pdf file to show the overlap of
EIF correlating genes identify from RNAseq of \code{tissue_type}

(2) bar graphs on screen and as pdf file to show the numbers of
identified correlating genes for each EIF4F subunit identify from
RNAseq of \code{tissue_type}

(3) heatmap on screen and as pdf file to show correlation strength
and clustering pattern of EIF4F correlating genes

(4) dotplot on screen and as pdf file to show the enriched pathways
in the genes from different clusters in the heatmap
}
\examples{
\dontrun{
.plot_Corr_RNAseq_TCGA_GTEX(x = "All")
.plot_Corr_RNAseq_TCGA_GTEX(x = "Lung")
}

}
\concept{composite function to call correlation analysis and plotting}
\keyword{internal}
