% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-Correlation.R
\name{.plot_Corr_RNAseq_TCGA_GTEX}
\alias{.plot_Corr_RNAseq_TCGA_GTEX}
\title{Perform correlation analysis on RNAseq data from all tumors and healthy tissues}
\usage{
.plot_Corr_RNAseq_TCGA_GTEX(x)
}
\arguments{
\item{x}{all tissue/tumor types or one specific type}
}
\value{
Venn digram, bargraph, heatmap, and dotplot
}
\description{
This function find correlating genes of EIF4F from
\code{TCGA_GTEX_RNAseq_sampletype} RNAseq data generated by \code{\link{initialize_RNAseq_data}}.

It identifies positively or negatively correlating genes from tumor or healthy samples,
find the overlap with Vennplot, compare their number with bargraph,
visualize the correlation strength with heatmap, and found enriched pathways within correlating genes

This function should not be used directly, only inside \code{\link{EIF4F_Corrgene_analysis}} function.
}
\examples{
\dontrun{
.plot_Corr_RNAseq_TCGA_GTEX(x = "All")
.plot_Corr_RNAseq_TCGA_GTEX(x = "Lung")
}

}
\keyword{internal}
