% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{.plot_KM_RNAseq_TCGA}
\alias{.plot_KM_RNAseq_TCGA}
\title{Survival analyses of TCGA patients with expression in their tumors by
Kaplan Meier method}
\usage{
.plot_KM_RNAseq_TCGA(gene_name, cutoff, tumor)
}
\arguments{
\item{gene_name}{gene name}

\item{cutoff}{percentage of gene expression for patient stratification}

\item{tumor}{all tumor types or specific type}
}
\description{
This function generates a Kaplan Meier plot to compare the expression of
one gene in TCGA cancer types.
}
\details{
This function
\itemize{
\item selects RNAseq of the query gene, survival data and cancer types from
the dataset \code{TCGA_RNAseq_OS_sampletype} prepared from
\code{\link[=initialize_survival_data]{initialize_survival_data()}}.
\item compares the survival data from patients with top or bottom percents of
gene expression, and plot the results as a KM curve plot by \code{\link[=.KM_curve]{.KM_curve()}}.
}

This function is not accessible to the user and will not show at the users'
workspace. It can only be called by the exported \code{\link[=EIF4F_Survival_analysis]{EIF4F_Survival_analysis()}}
function.

Side effects:

(1) KM curve plots on screen and as pdf files to correlate patient survival
probability with expression of \code{gene_name} in their tumors
}
\examples{
\dontrun{
plot.km.EIF.tumor(gene_name = "EIF4E", cutoff = 0.2,
tumor = "lung adenocarcinoma")
}

}
\seealso{
Other composite function to call survival analysis and plotting: 
\code{\link{.plot_CoxPH_RNAseq_TCGA}()}
}
\concept{composite function to call survival analysis and plotting}
\keyword{internal}
