% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{.plot_matrix_CNVcorr_TCGA}
\alias{.plot_matrix_CNVcorr_TCGA}
\title{Correlation matrix for CNV values}
\usage{
.plot_matrix_CNVcorr_TCGA(gene_list)
}
\arguments{
\item{gene_list}{gene names in a vector of characters}
}
\description{
generates correlation matrix for eIF4F CNV in tumors
from all TCGA cancer type combined
}
\details{
This function
\itemize{
\item selects a subset of unthreshold CNV values of only EIF4F gene
from the data frame \code{TCGA_CNV_value} - a global variable generated from \code{\link[=initialize_cnv_data]{initialize_cnv_data()}}.
\item uses the subset data to calculate the correlation coefficients
between every two genes and plot the correlation matrix with the function
\code{\link[=.matrix_plot]{.matrix_plot()}}
}

This function is not accessible to the user and will not show at the users'
workspace. It can only be called by the exported \code{\link[=EIF4F_CNV_analysis]{EIF4F_CNV_analysis()}} function.

side effect: the correlation matrix plot on screen and as saved pdf files
to show co-occurrence of \code{gene_list} CNV status in TCGA tumors
}
\examples{
\dontrun{
plot_matrix_CNVcorr_TCGA(c("EIF4A1", "EIF4E", "EIF4EBP1", "EIF4G1"))
}

}
\concept{Composite function to call CNV data analysis and plotting}
\keyword{internal}
