% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.plot_PCA_TCGA_GTEX}
\alias{.plot_PCA_TCGA_GTEX}
\title{\subsection{Master functions to call PCA and plot results}{

Perform PCA on RNAseq data from all tumors and healthy tissues
}}
\usage{
.plot_PCA_TCGA_GTEX(EIF.list)
}
\arguments{
\item{EIF.list}{gene names in a vector of characters}
}
\value{
PCA biplot, matrix, scree plot
}
\description{
This function performs standard PCA on selected RNAseq data from
\code{EIF.list} genes generated by \code{\link[=.get_df_subset]{.get_df_subset()}}.

It perform PCA on tumors from all TCGA cancer types,
PCA on healthy tissues from all GTEx healthy tissue types,
and PCA on combined TCGA tumors and GTEx healthy tissues.

This function should not be used directly, only inside \code{\link[=EIF4F_PCA]{EIF4F_PCA()}} function.
}
\examples{
\dontrun{
.plot_PCA_TCGA_GTEX(c(
  "EIF4E", "EIF4G1", "EIF4A1",
  "EIF4EBP1", "PABPC1", "MKNK1", "MKNK2"
))
}

}
\keyword{internal}
