% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.plot_PCA_TCGA_GTEX_tumor}
\alias{.plot_PCA_TCGA_GTEX_tumor}
\title{Perform PCA on RNAseq data from one tumor type and its matched healthy
tissue}
\usage{
.plot_PCA_TCGA_GTEX_tumor(gene_list, sample_type)
}
\arguments{
\item{gene_list}{gene names in a vector of characters}

\item{sample_type}{tissue type}
}
\description{
A composite function performs PCA on RNAseq data from one
tumor type and its matched healthy tissue, and plot results
}
\details{
This function
\itemize{
\item selects RNAseq data of \code{gene_list} genes in TCGA tumors and GTEx healthy
tissues with the same \code{tissue} of origin from \code{TCGA_GTEX_RNAseq_sampletype},
by \code{\link[=.get_df_subset]{.get_df_subset()}}.
\item performs PCA on combined tumor and healthy samples by \code{\link[=.RNAseq_PCA]{.RNAseq_PCA()}},
\item plots results as biplot, scree and matrix plots by \code{\link[=.biplot]{.biplot()}}
}

This function is not accessible to the user and will not show at the users'
workspace. It can only be called by the exported \code{\link[=EIF4F_PCA]{EIF4F_PCA()}} function.

Side effects:

(1) PCA biplots (PCA score plot + loading plot) on screen and as pdf files:
PCA score plot shows the clusters of samples based on their similarity and
loading plot shows how strongly each characteristic influences a principal
component.

(2) matrix plots on screen and as pdf files to show the quality of
representation of the variables.

(3) scree plots on screen and as pdf files to display how much variation
each principal component captures from the data.
}
\examples{
\dontrun{
.plot_PCA_TCGA_GTEX_tumor(c(
  "EIF4G1", "EIF4A1", "EIF4E",
  "EIF4EBP1", "PABPC1", "MKNK1", "MKNK2"), "Lung")
}

}
\seealso{
Other composite function to call PCA and plot results: 
\code{\link{.plot_PCA_CPTAC_LUAD}()},
\code{\link{.plot_PCA_TCGA_GTEX}()}
}
\concept{composite function to call PCA and plot results}
\keyword{internal}
