% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAProCorr.R
\name{.plot_scatter_RNApro_CCLE}
\alias{.plot_scatter_RNApro_CCLE}
\title{Correlation between CCLE RNAseq and proteomics data}
\usage{
.plot_scatter_RNApro_CCLE(gene_name)
}
\arguments{
\item{gene_name}{gene name}
}
\description{
A composite function generates correlation scatter plot for eIF4F RNAseq and
proteomics data from CCLE.
}
\details{
This function
\itemize{
\item merges the CCLE RNAseq values from EIF4F genes in the data frame
prepared from \code{\link[=.get_CCLE_RNAseq_subset]{.get_CCLE_RNAseq_subset()}} and proteomics data of the same
protein in the data frame prepared from \code{\link[=.get_CCLE_Proteomics_subset]{.get_CCLE_Proteomics_subset()}}.
\item uses the combined data to calculate the correlation coefficients
between protein and RNA levels, and plot the result with the function
\code{\link[=.RNApro_scatterplot]{.RNApro_scatterplot()}}
}

This function should not be used directly, only inside
\code{\link[=EIF4F_RNA_pro_correlation]{EIF4F_RNA_pro_correlation()}} function.

Side effects:

(1) scatter plot to show correlation between RNA and protein of
input \code{gene_name} expressions in CCLE samples
}
\examples{
\dontrun{
lapply(c("EIF4G1", "EIF4A1", "EIF4E"), .plot_scatter_RNApro_CCLE)
}

}
\seealso{
Other composite function to call RNA protein correlation analysis and
plotting: 
\code{\link{.plot_scatter_RNApro_LUAD}()}
}
\concept{composite function to call RNA protein correlation analysis and
plotting}
\keyword{internal}
