% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteomics.R
\name{.plot_scatterplot_protein_CPTAC}
\alias{.plot_scatterplot_protein_CPTAC}
\title{protein-protein coexpression in CPTAC LUAD}
\usage{
.plot_scatterplot_protein_CPTAC()
}
\description{
A composite function for coexpression and differential
expression analysis
}
\details{
This function is not accessible to the user and will not show at the users'
workspace. It can only be called by the exported \code{\link[=EIF4F_Proteomics_analysis]{EIF4F_Proteomics_analysis()}}
function.

Side effects:

(1) scatter plots to show correlation between two protein
expressions
}
\seealso{
Other composite function to call protein coexpression and differential
expression analysis: 
\code{\link{.plot_boxgraph_protein_CPTAC}()}
}
\concept{composite function to call protein coexpression and differential
expression analysis}
\keyword{internal}
