% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-Proteincorr.R
\name{.protein_boxplot}
\alias{.protein_boxplot}
\title{Boxplots of phosphor-proteomics data across clinic stages}
\usage{
.protein_boxplot(df, x)
}
\arguments{
\item{df}{a data frame of the combined proteomics and prosphorylation data
generated inside \code{\link{.plot_boxgraph_protein_CPTAC}}}

\item{x}{protein name}
}
\description{
This function should not be used directly,
only inside \code{\link{.plot_boxgraph_protein_CPTAC}}.
}
\keyword{internal}
