% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-PCA.R
\name{.protein_imputePCA}
\alias{.protein_imputePCA}
\title{Perform imputPCA on proteomics data}
\usage{
.protein_imputePCA(df, x)
}
\arguments{
\item{df}{\code{CPTAC.LUAD.Proteomics.Sample.subset} generated inside \code{\link{.plot_PCA_CPTAC_LUAD}}}

\item{x}{maximum number of components kept in the final results}
}
\value{
imputed PCA results
}
\description{
This function perform impute PCA on data generated by \code{\link{.get_df_subset}}.
It should not be used directly, only inside \code{\link{.plot_PCA_CPTAC_LUAD}} function.
}
\examples{
\dontrun{
.protein_imputePCA(CPTAC.LUAD.Proteomics.Sample.subset)
}
}
\keyword{internal}
