% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteincorr.R
\name{.protein_scatterplot}
\alias{.protein_scatterplot}
\title{Scatter plots of protein coexpression}
\usage{
.protein_scatterplot(df, x, y, z)
}
\arguments{
\item{x}{protein name}

\item{y}{protein name}

\item{z}{color scheme for the scatter plot}

\item{data}{input LUAD dataset generated from \code{\link[=.plot_scatterplot_protein_LUAD]{.plot_scatterplot_protein_LUAD()}}.}
}
\description{
This function should not be used directly, only inside \code{\link[=.plot_scatterplot_protein_LUAD]{.plot_scatterplot_protein_LUAD()}} function.
}
\keyword{internal}
