% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProRNACorrelation.R
\name{.RNApro_scatterplot}
\alias{.RNApro_scatterplot}
\title{Scatter plots of correlation between RNAseq and proteomics data}
\usage{
.RNApro_scatterplot(df, x, y)
}
\arguments{
\item{df}{dataframe of both RNAseq and proteomics generated inside}

\item{x}{protein or gene name}

\item{y}{database name. \code{CCLE} or \code{LUAD}}
}
\description{
This function should not be used directly, only inside
\code{\link[=.plot_scatter_RNApro_CCLE]{.plot_scatter_RNApro_CCLE()}} or \code{\link[=.plot_scatter_RNApro_LUAD]{.plot_scatter_RNApro_LUAD()}} function.
}
\keyword{internal}
