% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DEG.R
\name{.RNAratio_boxplot}
\alias{.RNAratio_boxplot}
\title{Box plots of differential RNA ratios across tumors}
\usage{
.RNAratio_boxplot(df, dashline, ylimit, filename)
}
\arguments{
\item{data}{output dataset generated from \code{\link{.RNAratio_selection}} function}
}
\description{
This function should not be used directly, only inside \code{\link{.plot_boxgraph_RNAratio_TCGA}}function.
}
\keyword{internal}
