% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{.RNAratio_selection}
\alias{.RNAratio_selection}
\title{Select RNA ratio data for plotting}
\usage{
.RNAratio_selection(df, x)
}
\arguments{
\item{df}{output dataset generated from \code{\link[=.RNAratio_calculation]{.RNAratio_calculation()}} function}

\item{x}{input argument for selected variables generated
from \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}} function}
}
\description{
This function should not be used directly,
only inside \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}}function.
}
\keyword{internal}
