% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DEG.R
\name{.RNAseq_boxplot}
\alias{.RNAseq_boxplot}
\title{Box plots of differential gene expression across tumors}
\usage{
.RNAseq_boxplot(df)
}
\arguments{
\item{data}{output dataset generated from \code{\link{.RNAseq_ind_gene}} function}
}
\description{
This function should not be used directly, only inside \code{\link{.plot_boxgraph_RNAseq_TCGA}}function.
}
\keyword{internal}
