% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{.RNAseq_grouped_boxplot}
\alias{.RNAseq_grouped_boxplot}
\title{Grouped box plots of RNA expression across tumors}
\usage{
.RNAseq_grouped_boxplot(df)
}
\arguments{
\item{data}{output dataset generated from \code{\link[=.RNAseq_all_gene]{.RNAseq_all_gene()}} function}
}
\description{
This function should not be used directly,
only inside \code{\link[=.plot_boxgraph_RNAseq_TCGA]{.plot_boxgraph_RNAseq_TCGA()}}function.
}
\keyword{internal}
