% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DEG.R
\name{.RNAseq_ind_gene}
\alias{.RNAseq_ind_gene}
\title{Analyzes differential gene expression in tumors vs adjacent normal tissues}
\usage{
.RNAseq_ind_gene(df, x)
}
\arguments{
\item{df}{\code{.TCGA_GTEX_RNAseq_sampletype_subset} generated inside \code{\link{.plot_boxgraph_RNAseq_TCGA}}}

\item{x}{one gene from the input argument of \code{\link{.plot_boxgraph_RNAseq_TCGA}}}
}
\value{
a data frame of differential gene expression in tumors vs adjacent normal tissues from individual TCGA cancer types
}
\description{
This function selects the RNAseq data from TCGA samples including all
tumors and solid tissue normal samples for comparison.

It should not be used directly, only inside \code{\link{.plot_boxgraph_RNAseq_TCGA}} function.
}
\keyword{internal}
