% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.RNAseq_PCA}
\alias{.RNAseq_PCA}
\title{Perform PCA on RNAseq data}
\usage{
.RNAseq_PCA(df, number_of_dimension)
}
\arguments{
\item{df}{\code{df} generated inside \code{\link[=.get_df_subset]{.get_df_subset()}}}

\item{number_of_dimension}{number of components kept in the final results}
}
\value{
PCA result
}
\description{
A helper function performs PCA on data generated by \code{\link[=.get_df_subset]{.get_df_subset()}}.
It should not be used directly, only inside \code{\link[=.plot_PCA_TCGA_GTEX]{.plot_PCA_TCGA_GTEX()}} or
\code{\link[=.plot_PCA_TCGA_GTEX_tumor]{.plot_PCA_TCGA_GTEX_tumor()}} function.
}
\examples{
\dontrun{
.RNAseq_PCA(df[[1]], 10)
}

}
\seealso{
Other helper function for PCA: 
\code{\link{.protein_imputePCA}()}
}
\concept{helper function for PCA}
\keyword{internal}
