% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DEG.R
\name{.RNAseq_tumortype}
\alias{.RNAseq_tumortype}
\title{Analyzes differential gene expression in primary, metastatic tumors vs adjacent normal tissues from all TCGA cancer types combined}
\usage{
.RNAseq_tumortype(df)
}
\arguments{
\item{df}{\code{.TCGA_GTEX_RNAseq_sampletype_subset} generated inside \code{\link{.plot_boxgraph_RNAseq_TCGA}}}
}
\value{
a data frame of differential gene expression in tumors vs adjacent normal tissues in individual TCGA cancer types
}
\description{
This function selects the RNAseq data from TCGA samples including
tumor samples that are labeled as “metastatic”, “primary tumor”, and “solid tissue normal” for comparison..

It should not be used directly, only inside \code{\link{.plot_boxgraph_RNAseq_TCGA}} function.
}
\keyword{internal}
