% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{.selected_biplot}
\alias{.selected_biplot}
\title{Plot subgroups of PCA results as biplots}
\usage{
.selected_biplot(res.pca, df, x, y, color)
}
\arguments{
\item{res.pca}{PCA results from from TCGA and GTEX combined RNAseq data}

\item{df}{combined RNAseq data with sample type annotation}

\item{x}{sample types for plot labeling}

\item{y}{type of primary diseases or healthy tissues for color of individuals}

\item{color}{color scheme used for individual sample on the PCA}

\item{folder}{sub directory name to store the output files}
}
\value{
PCA bioplot
}
\description{
This function draw biplot for PCA results
from TCGA and GTEX combined data from
\code{.get_df_subset(.TCGA_GTEX_sampletype_subset, "All")}.
}
\examples{
\dontrun{
.selected_biplot(
  res.pca = .RNAseq_PCA(df[[1]], 10), df = df[[2]],
  x = "Healthy Tissue (GTEx)", y = "sample.type", color = "#D55E00"
)
.selected_biplot(
  res.pca = .RNAseq_PCA(df[[1]], 10), df = df[[2]],
  x = "Healthy Tissue (GTEx)", y = "primary.site", color = col_vector
)
}

}
\keyword{internal}
