% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{.violinplot}
\alias{.violinplot}
\title{Violin plots of differential gene expression/ratio in primary,
metastatic tumors vs adjacent normal tissues}
\usage{
.violinplot(df, y.axis.title, y.axis.break, yintercept)
}
\arguments{
\item{data}{output dataset generated from \code{\link[=.RNAseq_tumortype]{.RNAseq_tumortype()}} or
\code{\link[=.RNAratio_tumortype]{.RNAratio_tumortype()}} function}
}
\description{
This function should not be used directly, only inside
\code{\link[=.plot_boxgraph_RNAseq_TCGA]{.plot_boxgraph_RNAseq_TCGA()}} or \code{\link[=.plot_boxgraph_RNAratio_TCGA]{.plot_boxgraph_RNAratio_TCGA()}} function.
}
\keyword{internal}
