% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eIF4F.analysis-package.R
\docType{package}
\name{eIF4F.analysis-package}
\alias{eIF4F.analysis}
\alias{eIF4F.analysis-package}
\title{eIF4F.analysis: A R package to analyze eIF4F subunits in human cancers}
\description{
This project contains data and source code generated by Dr. Su Wu, Ph.D., as part of her academic research work at Harvard Medical School.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/a3609640/eIF4F.analysis}
  \item Report bugs at \url{https://github.com/a3609640/eIF4F.analysis/issues}
}

}
\keyword{internal}
