% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{EIF4F_CNV_analysis}
\alias{EIF4F_CNV_analysis}
\title{\subsection{wrapper function to call all master functions with inputs}{

Perform all CNV related analysis and generate plots
}}
\usage{
EIF4F_CNV_analysis()
}
\value{
CNV analysis plots
}
\description{
A wrapper function to call all master functions for CNV analysis with inputs
}
\details{
This function run three master functions together with inputs:
\itemize{
\item \code{\link[=.plot_bargraph_CNV_TCGA]{.plot_bargraph_CNV_TCGA()}}
\item \code{\link[=.plot_matrix_CNVcorr_TCGA]{.plot_matrix_CNVcorr_TCGA()}}
\item \code{\link[=.plot_boxgraph_CNVratio_TCGA]{.plot_boxgraph_CNVratio_TCGA()}}
}
}
\examples{
\dontrun{
EIF4F_CNV_analysis()
}

}
