% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-Correlation.R
\name{EIF4F_Corrgene_analysis}
\alias{EIF4F_Corrgene_analysis}
\title{Perform PCA and generate plots}
\usage{
EIF4F_Corrgene_analysis()
}
\value{
correlating gene analysis plots
}
\description{
A wrapper function to call all master functions for PCA with inputs
}
\details{
This function run the master functions  \code{\link{.plot_Corr_RNAseq_TCGA_GTEX}}
with two inputs.
}
\examples{
\dontrun{
EIF4F_Corrgene_analysis()
}
}
