% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DEG.R
\name{EIF4F_DEG_analysis}
\alias{EIF4F_DEG_analysis}
\title{Perform all CNV related analysis and generate plots}
\usage{
EIF4F_DEG_analysis()
}
\value{
DEG analysis plots
}
\description{
A wrapper function to call all master functions for CNV analysis with inputs
}
\details{
This function run three master functions together:
\code{\link{.plot_boxgraph_RNAseq_TCGA}} and
\code{\link{.plot_boxgraph_RNAratio_TCGA}} with inputs
}
\examples{
\dontrun{
EIF4F_DEG_analysis()
}
}
