% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{EIF4F_PCA}
\alias{EIF4F_PCA}
\title{\subsection{Wrapper function to call all master functions with inputs}{

Perform PCA and generate plots
}}
\usage{
EIF4F_PCA()
}
\value{
CNV analysis plots
}
\description{
A wrapper function to call all master functions for PCA
}
\details{
This function run three master functions together:
\itemize{
\item \code{\link[=.plot_PCA_TCGA_GTEX]{.plot_PCA_TCGA_GTEX()}}
\item \code{\link[=.plot_PCA_TCGA_GTEX_tumor]{.plot_PCA_TCGA_GTEX_tumor()}}
\item \code{\link[=.plot_PCA_CPTAC_LUAD]{.plot_PCA_CPTAC_LUAD()}}
}
}
\examples{
\dontrun{
EIF4F_PCA()
}

}
