% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-PCA.R
\name{EIF4F_PCA}
\alias{EIF4F_PCA}
\title{Perform PCA and generate plots}
\usage{
EIF4F_PCA()
}
\value{
CNV analysis plots
}
\description{
A wrapper function to call all master functions for PCA with inputs
}
\details{
This function run three master functions together:
\code{\link{.plot_PCA_TCGA_GTEX}},
\code{\link{.plot_PCA_TCGA_GTEX_tumor}} and
\code{\link{.plot_PCA_CPTAC_LUAD}} with inputs
}
\examples{
\dontrun{
EIF4F_PCA()
}
}
