% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proteomics.R
\name{EIF4F_Proteomics_analysis}
\alias{EIF4F_Proteomics_analysis}
\title{Analyze co-expression among EIF4F subunits and differential expression}
\usage{
EIF4F_Proteomics_analysis()
}
\description{
A wrapper function to call two composite functions for protein
coexpression and differential expression.
}
\details{
This function run the composite functions \code{\link[=.plot_scatterplot_protein_CPTAC]{.plot_scatterplot_protein_CPTAC()}}
and \code{\link[=.plot_boxgraph_protein_CPTAC]{.plot_boxgraph_protein_CPTAC()}} with EIF4F gene name as inputs.

Side effects:

(1) scatter plot to show correlation between two protein expressions in CPTAC
LUAD samples

(2) boxplot to show the different expression of protein across
different clinical tumor stages
}
\examples{
\dontrun{
EIF4F_Proteomics_analysis()
}

}
\seealso{
Other wrapper function to call all composite functions with inputs: 
\code{\link{EIF4F_CNV_analysis}()},
\code{\link{EIF4F_Corrgene_analysis}()},
\code{\link{EIF4F_DEG_analysis}()},
\code{\link{EIF4F_PCA}()},
\code{\link{EIF4F_RNA_pro_correlation}()},
\code{\link{EIF4F_Survival_analysis}()}
}
\concept{wrapper function to call all composite functions with inputs}
