% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-ProRNACorrelation.R
\name{EIF4F_RNA_pro_correlation}
\alias{EIF4F_RNA_pro_correlation}
\title{Perform PCA and generate plots}
\usage{
EIF4F_RNA_pro_correlation()
}
\value{
correlating gene analysis plots
}
\description{
A wrapper function to call all master functions for RNA and protein correlation with inputs.
}
\details{
This function run the master functions \code{\link{.plot_scatter_RNApro_CCLE}} and
\code{\link{.plot_scatter_RNApro_LUAD}} with EIF4F gene name as inputs.
}
\examples{
\dontrun{
EIF4F_RNA_pro_correlation()
}
}
