% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{EIF4F_Survival_analysis}
\alias{EIF4F_Survival_analysis}
\title{\subsection{Wrapper function to call all master functions with inputs}{

Perform all related survival analysis and generate plots
}}
\usage{
EIF4F_Survival_analysis()
}
\value{
Survival analysis plots
}
\description{
A wrapper function to call all master functions for survival analysis with
inputs
}
\details{
This function runs two master functions together with inputs:
\itemize{
\item \code{\link[=.plot_KM_RNAseq_TCGA]{.plot_KM_RNAseq_TCGA()}}
\item \code{\link[=.plot_CoxPH_RNAseq_TCGA]{.plot_CoxPH_RNAseq_TCGA()}}
}
}
\examples{
\dontrun{
EIF4F_Survival_analysis()
}

}
