% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-Survival.R
\name{EIF4F_Survival_analysis}
\alias{EIF4F_Survival_analysis}
\title{Perform all related survival analysis and generate plots}
\usage{
EIF4F_Survival_analysis()
}
\value{
Survival analysis plots
}
\description{
A wrapper function to call all master functions for survival analysis with inputs
}
\details{
This function run three master functions together:
\code{\link{.plot_KM_RNAseq_TCGA}},
\code{\link{.plot_boxgraph_CNVratio_TCGA}} with inputs
}
\examples{
\dontrun{
EIF4F_Survival_analysis()
}
}
