% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-CNV.R
\name{initialize.cnv.data}
\alias{initialize.cnv.data}
\title{Read all CNV related datasets from TCGA}
\usage{
initialize.cnv.data()
}
\value{
TCGA CNV related datasets: including threshold, value and ratio data
}
\description{
Read all CNV related datasets from TCGA
}
