% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{initialize_cnv_data}
\alias{initialize_cnv_data}
\title{Read all CNV related datasets from TCGA}
\usage{
initialize_cnv_data()
}
\description{
This function reads all CNV related datasets from TCGA and its side effects
are three global variables.
}
\details{
\itemize{
\item \code{TCGA_CNV_value}: the unthreshold CNV value data generated from
\code{\link[=.get_TCGA_CNV_value]{.get_TCGA_CNV_value()}}
\item \code{TCGA_CNV_sampletype}: the merged dataset from \code{.TCGA.CNV} that contains
the threshold CNV data generated from \code{\link[=.get_TCGA_CNV]{.get_TCGA_CNV()}},
and \code{.TCGA_sampletype} that contains the annotation data from
\code{TCGA_phenotype_denseDataOnlyDownload.tsv} with the selection of
\code{sample.type} and \code{primary.disease} columns. \verb{Solid Tissue Normal} samples
are excluded.
\item \code{TCGA_CNVratio_sampletype}: the merged dataset from \code{.TCGA_CNV_ratio}
that contains the CNV ratio data generated from \code{\link[=.get_TCGA_CNV_ratio]{.get_TCGA_CNV_ratio()}},
and \code{.TCGA_sampletype}.
}
}
\examples{
\dontrun{
 initialize_cnv_data()
 }

}
