% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{initialize_cnv_data}
\alias{initialize_cnv_data}
\title{Read all CNV related datasets from TCGA}
\usage{
initialize_cnv_data()
}
\description{
A wrapper function reads all CNV related datasets from TCGA and its side effects
are three global variables.
}
\details{
Side effects:

(1) \code{TCGA_CNV_value}: the unthreshold CNV value data generated as the output
\code{\link[=.get_TCGA_CNV_value]{.get_TCGA_CNV_value()}}, which imports the download dataset
\code{Gistic2_CopyNumber_Gistic2_all_data_by_genes}. \code{TCGA_CNV_value} was stored
as \code{TCGA_CNV_value.csv} in \verb{~/Documents/EIF_output/ProcessedData} folder.

(2) \code{TCGA_CNV_sampletype}: the merged dataset from
\itemize{
\item \code{.TCGA.CNV} that contains the threshold CNV data from the download dataset
\code{Gistic2_CopyNumber_Gistic2_all_thresholded.by_genes} and was generated
from \code{\link[=.get_TCGA_CNV]{.get_TCGA_CNV()}},
\item \code{.TCGA_sampletype} that contains the annotation data from
\code{TCGA_phenotype_denseDataOnlyDownload.tsv} with the selection of
\code{sample.type} and \code{primary.disease} columns. \verb{Solid Tissue Normal} samples
are excluded.
\code{TCGA_CNV_sampletype} was stored as \code{TCGA_CNV_sampletype.csv} in
\verb{~/Documents/EIF_output/ProcessedData} folder.
}

(3) \code{TCGA_CNVratio_sampletype}: the merged dataset from
\itemize{
\item \code{.TCGA_CNV_ratio} that contains the CNV ratio data from the download
dataset \code{broad.mit.edu_PANCAN_Genome_Wide_SNP_6_whitelisted.gene.xena} and
was generated by \code{\link[=.get_TCGA_CNV_ratio]{.get_TCGA_CNV_ratio()}},
\item \code{.TCGA_sampletype}
\code{TCGA_CNVratio_sampletype} was stored as \code{TCGA_CNVratio_sampletype.csv} in
\verb{~/Documents/EIF_output/ProcessedData} folder.
}
}
\examples{
\dontrun{
 initialize_cnv_data()
 }

}
\seealso{
Other wrapper function for data initialization: 
\code{\link{initialize_RNAseq_data}()},
\code{\link{initialize_data}()},
\code{\link{initialize_phosphoproteomics_data}()},
\code{\link{initialize_proteomics_data}()},
\code{\link{initialize_survival_data}()}
}
\concept{wrapper function for data initialization}
