% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{initialize_data}
\alias{initialize_data}
\title{A wrapper function reads datasets from the download data files}
\usage{
initialize_data()
}
\description{
A wrapper function runs the five data initialization functions
to load the download data files.
\itemize{
\item \code{\link[=initialize_cnv_data]{initialize_cnv_data()}}
\item \code{\link[=initialize_RNAseq_data]{initialize_RNAseq_data()}}
\item \code{\link[=initialize_survival_data]{initialize_survival_data()}}
\item \code{\link[=initialize_proteomics_data]{initialize_proteomics_data()}}
\item \code{\link[=initialize_phosphoproteomics_data]{initialize_phosphoproteomics_data()}}
}
}
\details{
Side effects:

(1) \code{TCGA_CNV_value}imports the download dataset
\code{Gistic2_CopyNumber_Gistic2_all_data_by_genes}

(2) \code{TCGA_CNV_sampletype} imports and merges datasets from
\code{Gistic2_CopyNumber_Gistic2_all_thresholded.by_genes} and
\code{TCGA_phenotype_denseDataOnlyDownload.tsv}

(3) \code{TCGA_CNVratio_sampletype} imports and merges datasets from
\code{broad.mit.edu_PANCAN_Genome_Wide_SNP_6_whitelisted.gene.xena} and
\code{TCGA_phenotype_denseDataOnlyDownload.tsv}

(4) \code{TCGA_GTEX_RNAseq_sampletype} imports and merges datasets from
\code{TcgaTargetGtex_RSEM_Hugo_norm_count} and \code{TcgaTargetGTEX_phenotype.txt}

(5) \code{TCGA_RNAseq_OS_sampletype} imports and merges datasets from
\code{EB++AdjustPANCAN_IlluminaHiSeq_RNASeqV2.geneExp.xena},
\code{Survival_SupplementalTable_S1_20171025_xena_sp} and
\code{TCGA_phenotype_denseDataOnlyDownload.tsv}

(6) \code{CCLE_RNAseq} imports the download dataset
\code{CCLE_expression_full.csv}

(7) \code{CCLE_Anno}imports the download dataset \code{sample_info.csv}

(8) \code{CCLE_Proteomics} imports the download dataset
\code{protein_quant_current_normalized.csv}

(9) \code{CPTAC_LUAD_Proteomics} imports the download dataset \code{Protein.xlsx}

(10) \code{CPTAC_LUAD_RNAseq}imports the download dataset \code{RNA.xlsx}

(11) \code{CPTAC_LUAD_Phos}imports the download dataset \code{Phos.xlsx}

(12) \code{CPTAC_LUAD_Clinic_Sampletype} imports and merges datasets from
\code{S046_BI_CPTAC3_LUAD_Discovery_Cohort_Clinical_Data_r1_May2019.xlsx} and
\code{S046_BI_CPTAC3_LUAD_Discovery_Cohort_Samples_r1_May2019.xlsx}
}
\examples{
\dontrun{
initialize_data()
}

}
\seealso{
Other wrapper function for data initialization: 
\code{\link{initialize_RNAseq_data}()},
\code{\link{initialize_cnv_data}()},
\code{\link{initialize_phosphoproteomics_data}()},
\code{\link{initialize_proteomics_data}()},
\code{\link{initialize_survival_data}()}
}
\concept{wrapper function for data initialization}
