% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{initialize_dir}
\alias{initialize_dir}
\title{Set output directories}
\usage{
initialize_dir()
}
\description{
A wrapper function sets up the output directories
}
\details{
Side effects:

(1) output directory for CNV analysis results
\itemize{
\item \verb{~/Documents/EIF_output/CNV}
}

(2) output directory for differential gene expression and ratio analysis
results
\itemize{
\item \verb{~/Documents/EIF_output/DEG}
}

(3) output directories for survival analysis results
\itemize{
\item \verb{~/Documents/EIF_output/Survival/KM}
\item \verb{~/Documents/EIF_output/Survival/CoxPH}
}

(4) output directory for PCA results
\itemize{
\item \verb{~/Documents/EIF_output/PCA/TCGA_tumor+GTEX_healthy}
\item \verb{~/Documents/EIF_output/PCA/TCGA_tumor}
\item \verb{~/Documents/EIF_output/PCA/GTEX_healthy}
\item \verb{~/Documents/EIF_output/PCA/matched_tumor_and_healthy}
\item \verb{~/Documents/EIF_output/PCA/LUAD}
}

(5) output directory for analysis results on EIF4F correlating genes
\itemize{
\item \verb{~/Documents/EIF_output/CORs}
}

(6) output directory for correlation analysis between RNA and protein levels
\itemize{
\item \verb{~/Documents/EIF_output/RNApro}
}

(7) output directory for co-expression analysis among EIF4F subunits and
their differential expression
\itemize{
\item \verb{~/Documents/EIF_output/Proteomics/COEXP}
\item \verb{~/Documents/EIF_output/Proteomics/DIFFEXP}
}

(8) output directory for processed datasets used as global variables
their differential expression
\itemize{
\item \verb{~/Documents/EIF_output/ProcessedData}
}
}
\examples{
\dontrun{
initialize_dir()
}

}
\concept{wrapper function for setting up output directories}
