% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-Proteincorr.R
\name{initialize_phosphoproteomics_data}
\alias{initialize_phosphoproteomics_data}
\title{Read all phosphoproteomics related datasets from CPTAC LUAD}
\usage{
initialize_phosphoproteomics_data()
}
\description{
This function reads all phosphoproteomics related datasets from CPTAC LUAD.
It generates two global variable.

CPTAC_LUAD_Phos: the phosphoproteomics data of CPTAC from \code{Phos.xlsx}

CPTAC_LUAD_Clinic_Sampletype: the merged dataset from .CPTAC_LUAD_Clinic,
the CPTAC clinical data from \code{S046_BI_CPTAC3_LUAD_Discovery_Cohort_Clinical_Data_r1_May2019.xlsx},
and .CPTAC_LUAD_sampletype, the CPTAC annotation data from \code{S046_BI_CPTAC3_LUAD_Discovery_Cohort_Samples_r1_May2019.xlsx}.
}
\examples{
\dontrun{
initialize_phosphoproteomics_data()
}

}
