% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-ProRNACorrelation.R
\name{initialize_proteomics_data}
\alias{initialize_proteomics_data}
\title{Read all proteomics related datasets from CCLE and CPTAC LUAD}
\usage{
initialize_proteomics_data()
}
\description{
This function reads all proteomics related datasets from CCLE and CPTAC LUAD. It generates five global variables.

CCLE_RNAseq: the RNAseq data of CCLE from \code{CCLE_expression_full.csv}

CCLE_Anno: the annotation data of CCLE from \code{sample_info.csv}

CCLE_Proteomics: the proteomics data of CCLE from \code{protein_quant_current_normalized.csv}

CPTAC_LUAD_Proteomics: the proteomics data with annotation of CPTAC LUAD from \code{Protein.xlsx}

CPTAC_LUAD_RNAseq: the RNAseq data of CPTAC LUAD from \code{RNA.xlsx}
}
\examples{
\dontrun{
initialize_proteomics_data()
}

}
