% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAProCorr.R
\name{initialize_proteomics_data}
\alias{initialize_proteomics_data}
\title{Read all proteomics related datasets from CCLE and CPTAC LUAD}
\usage{
initialize_proteomics_data()
}
\description{
A wrapper function reads all proteomics related datasets from CCLE and
CPTAC LUAD.
}
\details{
Side effects:

(1) \code{CCLE_RNAseq}: the RNAseq data of CCLE from the download dataset
\code{CCLE_expression_full.csv}

(2) \code{CCLE_Anno}: the annotation data of CCLE from \code{sample_info.csv}

(3) \code{CCLE_Proteomics}: the proteomics data of CCLE from
\code{protein_quant_current_normalized.csv}

(4) \code{CPTAC_LUAD_Proteomics}: the proteomics data with annotation of CPTAC
LUAD study from the download data file \code{Protein.xlsx}

(5) \code{CPTAC_LUAD_RNAseq}: the RNAseq data of CPTAC LUAD samples from
the download data file \code{RNA.xlsx}
}
\examples{
\dontrun{
initialize_proteomics_data()
}

}
\seealso{
Other wrapper function for data initialization: 
\code{\link{initialize_RNAseq_data}()},
\code{\link{initialize_cnv_data}()},
\code{\link{initialize_data}()},
\code{\link{initialize_phosphoproteomics_data}()},
\code{\link{initialize_survival_data}()}
}
\concept{wrapper function for data initialization}
