% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DEG.R
\name{initialize_RNAseq_data}
\alias{initialize_RNAseq_data}
\title{Read all RNA-seq related datasets from TCGA and GTEx}
\usage{
initialize_RNAseq_data()
}
\description{
This function reads RNA-seq related datasets from TCGA and GTEx

.TCGA_GTEX_RNAseq: the recomputed RNAseq data from both TCGA and GTEx generated from \code{\link{.get_TCGA_GTEX_RNAseq}}

.TCGA_GTEX_sampletype: the annotation data from the "TcgaTargetGTEX_phenotype.txt"
dataset with selection of sample.type and primary.disease columns.

TCGA_GTEX_RNAseq_sampletype: the merged dataset from .TCGA_GTEX_RNAseq and .TCGA_GTEX_sampletype.
}
\examples{
\dontrun{
initialize_RNAseq_data()
}

}
