% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{initialize_survival_data}
\alias{initialize_survival_data}
\title{Read RNA-seq and survival datasets from TCGA}
\usage{
initialize_survival_data()
}
\description{
This function reads RNA-seq and survival datasets from TCGA.

side effect: \code{TCGA_RNAseq_OS_sampletype} the merged dataset from
\code{.TCGA_RNAseq}, \code{.TCGA_OS} and \code{.TCGA_sampletype}
\itemize{
\item \code{.TCGA_RNAseq}: the RNAseq data from TCGA generated from
\code{\link[=.get_TCGA_RNAseq]{.get_TCGA_RNAseq()}}
\item \code{.TCGA_OS}: the annotation data from the
\code{Survival_SupplementalTable_S1_20171025_xena_sp} dataset with selection of
\code{OS} and \code{OS.time} columns.
\item \code{.TCGA_sampletype}: the annotation data from the
\code{TCGA_phenotype_denseDataOnlyDownload.tsv} dataset with selection of
\code{sample.type} and \code{primary.disease} columns.
}
}
\examples{
\dontrun{
initialize_survival_data()
}

}
