% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{initialize_survival_data}
\alias{initialize_survival_data}
\title{Read RNA-seq and survival datasets from TCGA}
\usage{
initialize_survival_data()
}
\description{
A wrapper function to read RNA-seq and survival datasets from TCGA.

Side effects:

(1) \code{TCGA_RNAseq_OS_sampletype} a merged dataset from
\code{.TCGA_RNAseq}, \code{.TCGA_OS} and \code{.TCGA_sampletype}.
\itemize{
\item \code{.TCGA_RNAseq}: the RNAseq data from TCGA generated from
\code{\link[=.get_TCGA_RNAseq]{.get_TCGA_RNAseq()}}, which imports the dataset
\code{EB++AdjustPANCAN_IlluminaHiSeq_RNASeqV2.geneExp.xena}.
\item \code{.TCGA.OS}: the clinical data from
\code{Survival_SupplementalTable_S1_20171025_xena_sp}. We select three columns:
\code{OS} for overall survival status, \code{OS.time} for overall survival time and
\code{sample} for sample ID of each patient.
\item \code{.TCGA_sampletype}: the annotation data from the
\code{TCGA_phenotype_denseDataOnlyDownload.tsv} dataset. We select two columns
\code{sample.type} that annotates malignant tissues, and \code{primary.disease}
that annotates cancer types for each sample.
}

Only malignant tissue (Solid normal tissues are excluded) are selected for
survival analysis
}
\examples{
\dontrun{
initialize_survival_data()
}

}
\seealso{
Other wrapper function for data initialization: 
\code{\link{initialize_RNAseq_data}()},
\code{\link{initialize_cnv_data}()},
\code{\link{initialize_data}()},
\code{\link{initialize_phosphoproteomics_data}()},
\code{\link{initialize_proteomics_data}()}
}
\concept{wrapper function for data initialization}
